==> Documents/1.manuscript/1.DO/3b.High_Dimensional_Mediation.Rmd <==
---
title: "High Dimensional Mediation"
author: Anna L Tyler
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: 
  html_document:
    code_folding: hide
    collapsed: no
    toc: yes
    toc_float: yes
---

# Introduction
This workflow performs high-dimensional mediation on the 
Attie 500. 

It uses pre-adjusted transcriptomic and phenotypic data
generated in 1.Tissue_Expression.Rmd and 2.Trait_Selection.Rmd

The transcriptomic data have been processed to remove 
predicted genes and pseudogenes. The transript values
have also been adjust to remove the effects of sex 
and DO generation. For a more detailed explanaion,
see 1.Tissue_Expression.Rmd.

The phenotypic data have been pared down to informative,
relatively non-redundant traits. Sex, DO wave, and generation
were all regressed out. A more detailed explanation can 
be found in 2.Trait_Selection.Rmd.

For this workflow there are several parameters to set.
    1. exp.name: will determine the directory in which the results are stored.
    2. merge.tissues: if TRUE, the transcription kernel is an average of 
        all independent tissue kernels.


```{r get_args}
rm(list = ls()) #clear out R environment
#args <- commandArgs(trailingOnly=T)
num.perm = 1000

args <- c(1,1,0,1,0) #merged tissues, mediation model
#args <- c(0,0,0,1,0) #sep tissues, mediation model
#args <- c(1,0,0,1,0) #merged tissues, mediation model


#args <- c(1,1,1,1,0) #merged tissues, local imputation, mediation mode, weight adjusted

merge.tissues <- as.logical(args[1]) #if TRUE tissue transcriptome kernels are averaged. Otherwise they are used separtely
weight.adjusted <- as.logical(args[3]) #if TRUE, body weight is adjusted out of the phenotypes and transcriptome before running mediation
complete.mediation <- as.logical(args[4]) #if TRUE use a mediation network. If FALSE use a reactive network
delete_previous <- as.logical(args[5]) #if TRUE, previous results are overwritten

remove.local.effects <- FALSE #if TRUE, the effects of local genotype are
                             #removed from gene expression before kernelizing.
                             #This allows us to specifically analyze the correlation 
                             #between distal heritability and trait relevance 

use.local.effects <- FALSE   #if TRUE, the imputed transcriptomes are used
                            #to generate the transcriptome kernel. Cannot
                            #be TRUE if remove.locale.effects is TRUE

if(remove.local.effects && use.local.effects){
    stop("Only one of remove.local.effects or use.local.effects can be TRUE.")
}


if(is.na(merge.tissues)){stop("Please set the parameters for this run.")}

#hdm options
merge.text.options <- c("tissue_sep", "tissue_together")
weight.options <- c("_", "weight_adjusted")
mediation.options <- c("reactive", "complete_mediation")

#permutation options
perm.mat <- "mediator" #either permute the mediator matrix,
#perm.mat <- "kin"      #or the kinshp matrix
if(perm.mat == "mediator"){perm.text = "Permuting the mediator matrix."}
if(perm.mat == "kin"){perm.text = "Permuting the kinshp matrix."}

merge.text <- merge.text.options[(as.numeric(merge.tissues)+1)]
weight.text <- weight.options[(as.numeric(weight.adjusted)+1)]
mediation.text <- mediation.options[(as.numeric(complete.mediation)+1)]

exp.name <- paste(merge.text, weight.text, mediation.text, sep = "-")

if(merge.tissues){
    max.tx <- 1
}else{
    max.tx <- 4
}

adjust_before_fitting = NULL
#adjust_before_fitting = "Model_Scores_0.RDS" #we can look for higher dimensions
                            #of effects by adjusting for previous fitted models.
                            #This parameter specifies the name of the fit model
                            #scores to use for the adjustment.
                            #set to NULL to fit the original model
if(!is.null(adjust_before_fitting)){
    adjust.file <- here("Results", "DO", "High_Dim_Med", exp.name, adjust_before_fitting)
    check.file <- file.exists(adjust.file)
    if(!check.file){stop("Error in parameter: adjust_before_fitting\nThe model must be fit before the adjustment.")}
    model_scores <- readRDS(adjust.file)
}

```

This run had the following settings:

* Tissue treatment: `r merge.text`
* Weight adjustment: `r weight.text`
* Mediation type: `r mediation.text`
* Permutation type: `r perm.text`


```{r load_code}
is.interactive = FALSE
#is.interactive = TRUE
library("here")

results.dir <- here("Results", "DO", "High_Dim_Med", exp.name)
if(remove.local.effects){
    results.dir <- gsub("High_Dim_Med", "High_Dim_Med_distal", results.dir)
}
if(use.local.effects){
    results.dir <- gsub("High_Dim_Med", "High_Dim_Med_local", results.dir)
}

if(!file.exists(results.dir)){dir.create(results.dir, recursive = TRUE)}

parsed.data.dir <- here("Results", "DO", "Data") #a results folder that holds parsed data that we don't want to mix in with downloaded data

figure.data.dir <- here("Results", "Paper", "Source_Data") #Nature Communications requires data for all figures. Keep it here

all.fun <- list.files(here("Code"), full.names = TRUE, pattern = ".R")
for(i in 1:length(all.fun)){source(all.fun[i])}

```

```{r load_libraries, message = FALSE, warning = FALSE, error = FALSE}
all.packages <- c("pheatmap", "qtl2", "gprofiler2", "Matrix", "RGCCA", 
    "bnstruct", "igraph", "corpcor", "cluster", "RColorBrewer",
    "grid", "wordcloud", "stringr", "hexbin", "vioplot", "easyPubMed")

load_libraries(all.packages)
```

# High-Dimensional Mediation

## Data

Data were generated through a collaboration between the Attie
lab at the University of Wisconsin, and The Jackson Laboratory.
Detailed descriptions of mouse handling, phenotype gathering,
and initial data processing can be found elsewhere. 

Here we focus use transcriptomic data from five tissues: 
adipose, heart, islet, liver, and skeletal muscle.

I downloaded the following data from the 
[DO QTL viewer](https://churchilllab.jax.org/qtlviewer/attie/DO500HFD)
On March 20, 2023.

1. Adipose.RDS - adipose tissue transcriptome
2. Heart.RDS - cardiac muscle transcriptome
3. Islet.RDS - pancreatic islet transcriptome
4. Liver.RDS - liver transcriptome
5. QTLViewer_Geno_V10.Rdata - Genotypes for mice
6. SkeletalMuscle.RDS - skeletal muscle transcriptome
7. Clinical_Phenotypes_V11.RDS - clinical phenotypes for the Attie 500

The initial results from Isabela are in 
Results > Isabelas_Results. They are also stored in a Dropbox 
folder that Gary shared with me. 

```{r read_transcriptomes}
#read in transcriptomes processed by 1.Tissue_Expression.Rmd
#these are rank Z normalized and adjusted, but not mean centered and scaled.
adj.expr <- readRDS(file.path(parsed.data.dir, "Tissue_Expression_Adjusted.RDS")) #generated by 1b.Tissue_Expression.Rmd
tissue.names <- names(adj.expr)
#tissue.cols <- c("orange", "#8dd3c7", "tan", "brown") #colors that are not related to the DO/CC colors
#tissue.cols  <- colors.from.values(1:length(tissue.names), split.at.vals = TRUE, #nicer colors
#    split.points = seq(0.5, 4.5, 1), 
#    col.scale = c("blue", "red", "green", "purple"), light.dark = "d")
tissue.cols <- as.matrix(read.delim(here("Data", "general", "tissue_colors.txt"), header = FALSE, row.names = 1))[,1]
#barplot(rep(1, length(tissue.cols)), col = tissue.cols, names = names(tissue.cols))

load(here("Data", "DO", "QTLViewer_Geno_V10.Rdata")) #ensembl.version, genoprobs, K, map, markers
covar <- readRDS(file.path(parsed.data.dir, "Clinical_Phenotype_Covariates.RDS")) #generated by 1b.Trait_Selection.Rmd
gene.tables <- readRDS(file.path(parsed.data.dir, "Gene_Tables.RDS")) #generated by 1a.Tissue_Expression.Rmd

#lod tables
tissue.data.file <- lapply(tissue.names, 
  function(x) get.files(here("Data", "DO"), want = x, 
    dont.want = "vivo", full.names = FALSE))
tissue.lod <- lapply(tissue.data.file, 
    function(x) readRDS(here("Data", "DO", x))$lod.peaks$additive)
names(tissue.lod) <- tissue.names
```

```{r marker_dist, eval = FALSE}

for(chr in 1:19){
    cat("Chr", chr, "\n")
    split.marker <- strsplit(dimnames(genoprobs[[chr]])[[3]], "_")
    n.markers <- length(split.marker)
    marker.pos <- sapply(split.marker, function(x) as.numeric(x[2]))
    consec.markers <- consec_pairs(marker.pos)
    marker.dist <- consec.markers[,2] - consec.markers[,1]
    #plot(sort((marker.dist)))
    #abline(h = 100*1000)

    marker.windows <- vector(mode = "list", length = n.markers)
    #window.width <- 100*1000 #100 kb
    #window.width <- 200*1000 #200 kb
    window.width <- 500*1000 #500 kb
    #window.width = 2e6 #2 Mb
    #how many markers are in each 100kb block
    for(i in 1:n.markers){
        this.pos <- marker.pos[i]
        min.pos <- this.pos - window.width
        max.pos <- this.pos + window.width
        above.min <- which(marker.pos >= min.pos)
        below.max <- which(marker.pos <= max.pos)
        in_window <- intersect(above.min, below.max)
        marker.windows[[i]] <- in_window
    }
    #num.window.markers <- sapply(marker.windows, length)
    #hist(num.window.markers)

    #get all unique pairs of markers at this distance
    all.pairs <- cbind(unlist(lapply(1:n.markers, function(x) rep(x, length(marker.windows[[x]])))), unlist(marker.windows))
    sorted.pairs <- t(apply(all.pairs, 1, sort))
    u_pairs <- unique(sorted.pairs)
    pair.dist <- marker.pos[u_pairs[,2]] - marker.pos[u_pairs[,1]]

    dist.mat <- matrix(NA, nrow = n.markers, ncol = n.markers)
    dist.V <- rep(NA, nrow(u_pairs))
    for(i in 1:nrow(u_pairs)){
        report.progress(i, nrow(u_pairs))
        marker1 <- u_pairs[i,1]
        marker2 <- u_pairs[i,2]
        marker.cor <- cor(as.vector(genoprobs[[chr]][,,marker1]), 
            as.vector(genoprobs[[chr]][,,marker2]))
        dist.mat[marker1, marker2] <- marker.cor
        dist.V[i] <- marker.cor
    }

    dist.table <- cbind("pairwise_distance_bp" = pair.dist, 
        "pairwise_correlation_R" = dist.V)
    saveRDS(dist.table, here("Results", "For_Reviews", 
        paste0("LD_", window.width/1000, "kb_Chr", chr, ".RDS")))
    cat("\n")
}


pdf(here("Results", "For_Reviews", paste0("LD_", window.width/1000, "kb.pdf")), 
    width = 10, height = 5)
    for(chr in 1:19){
        dist.table <- readRDS(here("Results", "For_Reviews", 
            paste0("LD_", window.width/1000, "kb_Chr", chr, ".RDS")))
        pair.dist <- dist.table[,1]
        dist.V <- dist.table[,2]
        test <- bin_by_hist(pair.dist/1000, dist.V, 
            ylab = "Marker Correlation (R)",
            xlab = "Distance (kb)", ylim = c(0, 1.1), 
            main = paste("Chr", chr), 
            breaks = 15, error.type = "sd")
        abline(h = seq(0, 1, 0.1), lty = 2, col = "darkgray")
    }        
dev.off()
```

## Read clinical phenotypes

Read in clinical phenotypes. These were rank normalized, 
adjusted for sex and DO generation, and then mean centered
and standardized in 1a.Trait_Selection.Rmd

```{r pheno}
scaled.pheno <- readRDS(file.path(parsed.data.dir, "Clinical_Phenotypes_Adjusted.RDS")) #Generated by 1a.Trait_Selection.Rmd
```

## Kinship matrix

```{r germline_kin}

#calculate the germline regardless of which kinship 
#type we have specified. If we have specified the 
#functional kinship, we will compare it to the 
#germline kinship.

kin.file <- file.path(parsed.data.dir, "overall.kinship.RDS")
if(!file.exists(kin.file)){
    Kg = calc_kinship(genoprobs, "overall")
    saveRDS(Kg, kin.file)
}else{
    Kg <- readRDS(kin.file)
}

```

Subset and center kinship matrix based on DO generation.
Make separate matrices for each tissue, because each tissue
has different individuals in it.

```{r subset}
gen.col <- grep("Generation", colnames(covar))
do_gen <- lapply(adj.expr, function(x) covar[rownames(x),gen.col])
gen_cent = lapply(do_gen, function(x) x %*% Diagonal(dim(x)[2], 1 / colSums(x)) %*% t(x)) # Centering matrix

common.ind <- lapply(1:length(gen_cent), function(x) intersect(rownames(gen_cent[[x]]), rownames(Kg)))

#subset kinship matrix to match the individuals in each tissue.
#remove DO generation from kinship matrix
Kg_sub = lapply(1:length(gen_cent), function(x) Kg[common.ind[[x]], common.ind[[x]]])
cent_sub <- lapply(1:length(gen_cent), function(x) gen_cent[[x]][common.ind[[x]], common.ind[[x]]])
Kg_cent = lapply(1:length(gen_cent), function(x) Kg_sub[[x]] - cent_sub[[x]] %*% Kg_sub[[x]] - Kg_sub[[x]] %*% cent_sub[[x]] + cent_sub[[x]] %*% Kg_sub[[x]] %*% cent_sub[[x]])
```


## Process transcriptome

If specified, remove the effects of the
local eQTLs from each transcript before kernelizing. 
This allows us to look specifically at the relationship
between distal heritability and trait relevance without
comtamination from eQTLs. If we leave the effects of 
local eQTLs in, we allow all drivers of transcript
variation to potentially explain phenotypic variation.
To the extent that there are good TWAS hits, these
can be accounted for in this analysis.

After removing the local eQTL effects, mean center 
and standardize residual transcriptome. 

The phenotypes are already mean centered and scaled.

```{r remove_local_effects}

if(remove.local.effects){
    residual.expr.files <- sapply(tissue.names, function(x) file.path(parsed.data.dir, 
        paste0(x, "_Residual.RDS")))

    for(tx in 1:length(tissue.names)){
        if(!file.exists(residual.expr.files[tx])){
            tx.expr <- adj.expr[[tx]]
            resid.tx  <- matrix(NA, nrow = nrow(tx.expr), ncol = ncol(tx.expr))
            dimnames(resid.tx) <- dimnames(tx.expr)
            gene.info <- gene.tables[[tx]]
            for(ts in 8776:ncol(tx.expr)){
                report.progress(ts, ncol(tx.expr))
                nearest.marker <- gene.info$nearest.marker.id[which(gene.info$gene.id == colnames(tx.expr)[ts])]
                if(nearest.marker == ""){next()}
                nearest.geno <- pull_genoprobpos(genoprobs, marker = nearest.marker)
                model <- fit1(nearest.geno, tx.expr[,ts], kinship = Kg)
                resid.tx[,ts] <- model$resid
            }
            saveRDS(resid.tx, residual.expr.files[tx])
        }
    }
    residual.expr <- lapply(residual.expr.files, readRDS)

    #remove any transcripts that were skipped
    for(tx in 1:length(tissue.names)){
        na.idx <- which(apply(residual.expr[[tx]], 2, function(x) all(is.na(x))))
        to.keep <- setdiff(1:ncol(residual.expr[[tx]]), na.idx)
        residual.expr[[tx]] <- residual.expr[[tx]][,to.keep]
    }
}else{
residual.expr <- adj.expr
}
```

```{r use.local.effects}
#if we are going to use local effects to generate the transcriptome
if(use.local.effects){
    local.expr.files <- list.files(here("Results", "DO", "Imputed_Transcriptomes"), 
        pattern = "Local_Imputation", full.names = TRUE)
    residual.expr <- lapply(local.expr.files, readRDS)

    #remove transcripts for which we could not impute expression
    #(non-autosomal)
    for(tx in 1:length(residual.expr)){
        keep.cols <- which(!apply(residual.expr[[tx]], 2, function(x) all(is.na(x))))
        residual.expr[[tx]] <- residual.expr[[tx]][,keep.cols]
    }
}
```



```{r scale}

#use the residual expression to build the kernel
scaled.expr <- lapply(residual.expr, function(x) apply(x, 2, scale))
#add back the rownames that scaling removes
for(i in 1:length(scaled.expr)){
    rownames(scaled.expr[[i]]) <- rownames(adj.expr[[i]])
}
```

## Build transcriptomic kernel

```{r kernelize}
if(!is.null(adjust_before_fitting)){
    reg.expr <- lapply(scaled.expr, function(x) adjust(x, model_scores[[1]][,"Mediator",drop=F]))
}else{
    reg.expr <- scaled.expr
}
Kt <- lapply(reg.expr, function(x) x %*% t(x) / dim(x)[2])


if(merge.tissues){
    #make a single Kt matrix that is the average of all expression kernels
    #it is repeated for all tissues, so we don't need to change the structure
    #of the analysis
    #This will represent the structure of expression across all tissues
    common.ind <- Reduce("intersect", lapply(Kt, rownames))
    total.Kt <- Reduce("+", lapply(Kt, function(x) x[common.ind, common.ind]))
    Kt <- lapply(1:length(tissue.names), function(x) total.Kt/length(tissue.names))
    names(Kt) <- tissue.names
}
```

## Build phenotype kernel

The plot below shows where there are missing values in the
phenotype matrix. Red indicates missing, blue indicates a filled
cell. There are a few individuals that are missing a few entries. 
They are not concentrated in any particular generation or sex.

```{r pheno_missing}
common.ind <- intersect(rownames(Kt[[1]]), rownames(scaled.pheno))

#select a subset of phenotypes if desired
#sub.pheno <- scaled.pheno[,c("ins_final", "Glu_tAUC", "Ins_tAUC", "HOMA_B", "HOMA_IR")] #chr 11-related
#sub.pheno <- scaled.pheno[,c("weight_final", "food_ave", "TG_6", "TG_10", "TG_14", "TG_final", "Chol_PC1", "Chol_PC2", "Chol_PC3", "Chol_PC4", "leptin", "adiponectin", "fat_pad_weight", "adiposity")] #fat-related traits
#sub.pheno <- scaled.pheno[,c("TG_6wk", "TG_10wk", "TG_14wk")]
sub.pheno <- scaled.pheno[common.ind,]

pheno.missing <- matrix(0, nrow = nrow(sub.pheno), ncol = ncol(sub.pheno))
pheno.missing[which(is.na(sub.pheno))] <- 1
pheatmap(pheno.missing, main = "Missing Values")

missing.ind <- which(rowSums(pheno.missing) > 0)
ismissing <- rep(0, nrow(sub.pheno))
ismissing[missing.ind] <- 1

#par(mfrow = c(3,3))
pval <- rep(NA, ncol(covar))
for(i in 1:ncol(covar)){
    cont.table <- table(data.frame(cbind(ismissing, covar[common.ind,i])))
    #barplot(cont.table, beside = TRUE)
    sig.test <- fisher.test(cont.table)
    pval[i] <- sig.test$p.value
}
barplot(-log10(pval), names = colnames(covar), las = 2, ylab = "-log10(pval)",
main = "-log10(pval) for Fisher exact test\non missingness contingency table",
ylim = c(0,3))
```

```{r pheno_kernel}
#there are a few missing values. Impute these
sub.pheno <- knn.impute(sub.pheno)

if(!is.null(adjust_before_fitting)){
    reg.pheno <- adjust(sub.pheno, model_scores[[1]][,"Outcome", drop=FALSE])
}else{
    reg.pheno <- sub.pheno
}
Kp <- reg.pheno %*% t(reg.pheno) / dim(reg.pheno)[2]
```

## Perform high-dimensional mediation

```{r mediation}

if(!is.null(adjust_before_fitting)){
    adjust_numeral <- as.numeric(gsub(".RDS", "", strsplit(basename(adjust.file), "_")[[1]][3])) + 1
}else{
    adjust_numeral  <- 0
}

transcript_loading_file <- file.path(results.dir, paste0("Loadings_Transcripts_", adjust_numeral, ".RDS"))
trait_loading_file <- file.path(results.dir, paste0("Loadings_Traits_", adjust_numeral, ".RDS"))
marker_loading_file <- file.path(results.dir, paste0("Loadings_Markers_", adjust_numeral, ".RDS"))
model_file <- file.path(results.dir, paste0("Model_Scores_", adjust_numeral, ".RDS"))

if(!file.exists(transcript_loading_file) || delete_previous){
    transcript_loadings <- trait_loadings <- model_scores <- vector(mode = "list", length = length(tissue.names))
    names(transcript_loadings) <- names(trait_loadings) <- names(model_scores) <- tissue.names
    flags <- rep(NA, length(tissue.names))
    stopping.reasons <- rep(NA, length(tissue.names))
    for(tx in 1:length(tissue.names)){
        if(is.interactive){cat("\n", tissue.names[tx], "\n")}
        
        #assign matrices to proper positions
        if(complete.mediation){
            #transcripts are the mediator
            mediator.kernel = Kt[[tx]]; transcript.col <- "Mediator"
            #phenotypes are the outcome
            outcome.kernel = Kp; trait.col <- "Outcome"
        }else{
            #phenotypes are the mediator
            mediator.kernel = Kp;trait.col <- "Mediator" 
            #transcripts are the outcome
            outcome.kernel = Kt[[tx]]; transcript.col = "Outcome" 
        }
        
        fit.result <- high_dim_med(causal.matrix = Kg_cent[[tx]], 
            mediating.matrix = mediator.kernel, outcome.matrix = outcome.kernel, 
            min.weight.diff = 1e-5, max.iter = 15, scheme = "centroid", 
            verbose = is.interactive, kernel.c = TRUE, kernel.m = TRUE, 
            kernel.o = TRUE)

        #test <- scan1(genoprobs, fit.result[[1]][,2])
        #plot(test, map = map)

        model_scores[[tx]] <- fit.result[[1]]
        flags[tx] <- fit.result[[2]]
        stopping.reasons[tx] <- fit.result[[3]]

        #calculate the trait loadings. This is the 
        #correlation between the scaled phenotype
        #and the phenotype score found by the model
        trait_cor <- calc_loadings(scores = model_scores[[tx]][,trait.col,drop = FALSE], 
            data.mat = sub.pheno)
        

        #if the traits are mostly negatively correlated with 
        #the scores, multiply the model scores by -1
        #and recalculate
        if(mean(trait_cor) < 0){
            model_scores[[tx]] <- model_scores[[tx]] * -1
            trait_cor <- calc_loadings(model_scores[[tx]][,trait.col,drop = FALSE], 
                sub.pheno)
        }
        trait_loadings[[tx]] <- t(trait_cor)
    
        #calculate transcript loadings
        #This is the correlation between the scaled expression
        #and the transcript score found by the model
        tx_load <- calc_loadings(model_scores[[tx]][,transcript.col,drop = FALSE], 
            scaled.expr[[tx]])
        transcript_loadings[[tx]] <- t(tx_load)
        }

    saveRDS(transcript_loadings, transcript_loading_file)
    saveRDS(trait_loadings, trait_loading_file)
    saveRDS(model_scores, model_file)
}else{
    transcript_loadings <- readRDS(transcript_loading_file)
    trait_loadings <- readRDS(trait_loading_file)
    model_scores <- readRDS(model_file)
}


comp.pheno.herit <- sapply(model_scores, function(x) as.numeric(est_herit(x[,"Outcome"], Kg)))
comp.pheno.herit.error <- sapply(model_scores, function(x) as.numeric(est_herit_error(x[,"Outcome"], Kg)))
pheno.herit <- c("Heritability" = comp.pheno.herit[1], "Error" = comp.pheno.herit.error[1])
saveRDS(pheno.herit, here("Results", "Paper", "Composite_Trait_Heritability.RDS"))

pheno.decomp <- plot.decomp(reg.pheno, pc = ncol(reg.pheno), plot.results = FALSE)
pheno.pc <- pheno.decomp$u
pc.var.exp <- pheno.decomp$var.exp
common.id <- lapply(model_scores, function(x) intersect(rownames(reg.pheno), rownames(x)))
common.idx <- lapply(1:length(common.id), 
    function(x) cbind(match(common.id[[x]], rownames(reg.pheno)), match(common.id[[x]], rownames(model_scores[[x]]))))
pheno.model.cor <- lapply(1:length(model_scores), 
    function(x) apply(pheno.pc[common.idx[[x]][,1],], 2, 
    function(y) cor(model_scores[[x]][common.idx[[x]][,2],"Outcome"], y)))
names(pheno.model.cor) <- tissue.names
model.var.exp <- lapply(pheno.model.cor, function(x) sum(x*pc.var.exp))

saveRDS(lapply(model.var.exp, function(x) x*100), here("Results", "Paper", 
    paste0("Model_Variance_Explained_", basename(results.dir), ".RDS")))
```


```{r test_mediate, eval = FALSE}

#Does the liver expression of Pparg account for the heritability of plasma triglycerides?

gene.name <- "Slc27a4"
gene.name <- "Arpp19"
tissue <- "Liver"
tissue.idx <- which(tissue.names == tissue)

gene.id <- gene.tables[[tissue.idx]]$gene.id[which(gene.tables[[tissue.idx]]$symbol == gene.name)]
gene.expr <- scaled.expr[[tissue.idx]][,gene.id,drop=FALSE]

par(mfrow = c(5,4), mar = c(2,2,2,2))
for(tr in 1:ncol(scaled.pheno)){
    trait <- scaled.pheno[,tr,drop=FALSE]

    trait.herit <- as.numeric(est_herit(trait, Kg))

    #remove gene expression from trait
    no.expr <- adjust(trait, gene.expr)
    no.expr.herit <- as.numeric(est_herit(no.expr, Kg))

    barplot_with_num(signif(c(trait.herit, signif(no.expr.herit)), 2), 
        main = colnames(scaled.pheno)[tr], ylim = c(0, 0.6))
}

#no, the heritability of TG levels is not reduced after accounting for Pparg gene expression
#In contrast, adipose leptin expression strongly mediates genetic effects of fasting insulin levels


#what is the best expression mediator of TG in liver?
trait.name <- "TG_Fasting"
trait.idx <- which(colnames(scaled.pheno) == trait.name)
trait <- scaled.pheno[,trait.idx,drop=FALSE]
trait.herit <- as.numeric(est_herit(trait, Kg))

tissue <- "Liver"
tissue.idx <- which(tissue.names == tissue)

med.herit <- rep(NA, ncol(scaled.expr[[tissue.idx]]))
names(med.herit) <- colnames(scaled.expr[[tissue.idx]])
for(tr in 1:ncol(scaled.expr[[tissue.idx]])){
    report.progress(tr, ncol(scaled.expr[[tissue.idx]]))
    gene.expr <- scaled.expr[[tissue.idx]][,tr,drop=FALSE]
    #remove gene expression from trait
    no.expr <- adjust(trait, gene.expr)
    med.herit[tr] <- as.numeric(est_herit(no.expr, Kg))
}

herit.drop <- trait.herit-med.herit
plot(sort(herit.drop))
which.max(herit.drop)
```

## Causal Networks {.tabset .tabset-fade .tabset-pills}

The following plot shows the partial correlations between
the causal factor, the mediator, and the outcome. Line
width corresponds to partial correlation, and line color
corresponds with the sign of the partial correlation. Orange
is negative and green is positive.

If the tissues were run separately, there will be one plot
per tissue. Otherwise, there will be just one plot for all tissues.

```{r plot_nets, results = "asis", fig.width = 7, fig.height = 7}
#layout(get.layout.mat(max.tx))
for(tx in 1:max.tx){
 if(max.tx == 1){
        cat("### All Tissues\n")
        file.append <- "all_tissues"
        node.col = "lightblue"
    }else{
        cat("###", tissue.names[tx], "\n")
        file.append <- tissue.names[tx]
        node.col = tissue.cols[tx]
    }    
    tx.pcor <- pcor.shrink(model_scores[[tx]], verbose = FALSE)
    tx.pcor[lower.tri(tx.pcor, diag = TRUE)] <- 0
    net <- graph_from_adjacency_matrix(tx.pcor, weighted = TRUE, mode = "directed")
    e.weight <- E(net)$weight
    plot(net, edge.width = abs(e.weight*10), edge.color = sign(e.weight)+2, 
        layout = layout_on_grid, edge.label = signif(e.weight, 2),
        edge.label.cex = 1.5, vertex.size = 30, main = file.append,
        vertex.color = node.col)
    if(max.tx > 1){
        cat("\n\n")
    }

    saveRDS(tx.pcor, here("Results", "Paper", paste0("Pcor_", file.append, ".RDS")))
}
```

## Permutations {.tabset .tabset-fade .tabset-pills}

Run permutations to calulate a null distribution for the path
coefficient. This is the correlation between the causal matrix 
and the outcome matrix. To do this, we shuffle the names of the 
mediator kernel. 

This decorrelates the 


```{r perm}

perm.file <- file.path(results.dir, 
    paste0("Permuted_Stats_", perm.mat, "_permuted_", adjust_numeral, ".RDS")) #holds permuted path coefficients and partial correlations between components
null.mediator.file <- file.path(results.dir, 
    paste0("Permuted_Mediator_Scores_", perm.mat, "_permuted_", adjust_numeral, ".RDS")) #holds permuted mediator scores, which can be used to calculate null transcript loadings

if(!file.exists(perm.file) || delete_previous){

    all.perms <- vector(mode = "list", length = max.tx)
    null.mediator.scores <- vector(mode = "list", length = max.tx)

    for(tx in 1:max.tx){

        if(complete.mediation){
            #for the mediation model, the transcript kernel is the mediator
            mediator.kernel = Kt[[tx]]
            outcome.kernel = Kp
        }else{
            #for the reactive model, the phenotype is the mediator
            mediator.kernel = Kp
            outcome.kernel = Kt[[tx]]
        }
        
        common.names <- intersect(rownames(mediator.kernel), rownames(outcome.kernel))
        perm_stats <- matrix(NA, nrow = num.perm, ncol = 4)
        colnames(perm_stats) <- c("path_coef", "X-M", "M-Y", "X-Y")
        null.med.mat <- matrix(NA, nrow = length(common.names), ncol = num.perm)        
        rownames(null.med.mat) <- common.names

        for(p in 1:num.perm){

            if(is.interactive){report.progress(p, num.perm)}
            
            if(perm.mat == "mediator"){
                #permute the mediator
                perm_names <- sample(rownames(mediator.kernel))
                perm_mediator <- mediator.kernel
                rownames(perm_mediator) <- colnames(perm_mediator) <- perm_names
           
                perm_scores <- high_dim_med(causal.matrix = Kg_cent[[tx]], 
                    mediating.matrix = perm_mediator, outcome.matrix = outcome.kernel, 
                    min.weight.diff = 1e-5, max.iter = 15, scheme = "centroid", 
                    verbose = FALSE, kernel.c = TRUE, kernel.m = TRUE, 
                    kernel.o = TRUE)
                        
            }
            if(perm.mat == "kin"){
                #permute the mediator
                perm_names <- sample(rownames(Kg_cent[[tx]]))
                perm_kin <- Kg_cent[[tx]][perm_names, perm_names]
                rownames(perm_kin) <- colnames(perm_kin) <- rownames(Kg_cent[[tx]])

                perm_scores <- high_dim_med(causal.matrix = perm_kin, 
                    mediating.matrix = mediator.kernel, outcome.matrix = outcome.kernel, 
                    min.weight.diff = 1e-5, max.iter = 15, scheme = "centroid", 
                    verbose = FALSE, kernel.c = TRUE, kernel.m = TRUE, 
                    kernel.o = TRUE)
                
            }
        
            #keep mediator model scores
            null.med.mat[,p] <- perm_scores[[1]][,"Mediator"]
            #keep stats
            perm_stats[p,] <- path_coef(perm_scores[[1]])
        }

    all.perms[[tx]] <- perm_stats
    null.mediator.scores[[tx]] <- null.med.mat
    }

    saveRDS(all.perms, perm.file)
    saveRDS(null.mediator.scores, null.mediator.file)

}else{
    all.perms <- readRDS(perm.file)
    null.mediator.scores <- readRDS(null.mediator.file)
}

```

The following plots show null distributions of the path 
coefficient for each tissue, unless tissues were merged, 
and then only one set of null distribution is shown. 

The dot plot shows the relationship between the X-M 
correlation and the M-Y correlation. The permutations 
show that you can make one of them quite high, but 
getting both high is very difficult and can only be 
done with the true data.

The histogram shows the null distribution of the path
coefficient along with the observed path coeffient.
The observed path coefficient is well outside the 
null distribution.


```{r plot_perm, results = "asis", fig.height = 5, fig.width = 5}

for(tx in 1:max.tx){
    if(max.tx == 1){
        cat("### All Tissues\n")
        tissue.col = "darkgray"
    }else{
        cat("###", tissue.names[tx], "\n")
        tissue.col = tissue.cols[tx]
    }
    
    tissue.path.stats <- path_coef(model_scores[[tx]])

    #write out path stats for comparison in ms
    saveRDS(tissue.path.stats, file.path(results.dir, "Path_Stats.RDS"))

    tissue.path.coef <- tissue.path.stats[1]

    axis.max <- max(c(all.perms[[tx]][,c("X-M", "M-Y")], tissue.path.stats[c("X-M", "M-Y")]), na.rm = TRUE)
    axis.min <- min(c(all.perms[[tx]][,c("X-M", "M-Y")], tissue.path.stats[c("X-M", "M-Y")]), na.rm = TRUE)

    plot(all.perms[[tx]][,"X-M"], all.perms[[tx]][,"M-Y"], 
        xlim = c(axis.min, axis.max), 
        ylim = c(axis.min, axis.max), xlab = "X-M correlation", 
        ylab = "M-Y correlation", pch = 16, col = tissue.col)
    points(tissue.path.stats["X-M"], tissue.path.stats["M-Y"], 
        col = "red", pch = 16)
    abline(0,1)

    path.min <- min(c(all.perms[[tx]][,"path_coef"], tissue.path.coef), na.rm = TRUE)
    path.max <- max(c(all.perms[[tx]][,"path_coef"], tissue.path.coef), na.rm = TRUE)
    hist(all.perms[[tx]][,"path_coef"], xlim = c(path.min, path.max), 
        main = "Path Coef. Permutations", xlab = "", breaks = 100,
        col = tissue.col)
    abline(v = tissue.path.coef, col = "red")

    cat("\n\n")
}

calc.emp.p(tissue.path.coef, all.perms[[tx]][,"path_coef"])
```

```{r fig3B_C, eval = FALSE}
#old figure panels without local and distal kernels
pdf(here("Results", "Paper", "Fig3B_C.pdf"), width = 5, height = 5)
tx = 1

tissue.path.stats <- path_coef(model_scores[[tx]])
tissue.path.coef <- tissue.path.stats[1]

axis.max <- max(c(all.perms[[tx]][,c("X-M", "M-Y")], tissue.path.stats[c("X-M", "M-Y")]), na.rm = TRUE)
axis.min <- min(c(all.perms[[tx]][,c("X-M", "M-Y")], tissue.path.stats[c("X-M", "M-Y")]), na.rm = TRUE)

plot(all.perms[[tx]][,"X-M"], all.perms[[tx]][,"M-Y"], 
    xlim = c(axis.min, axis.max), 
    ylim = c(axis.min, axis.max), xlab = "X-M correlation", 
    ylab = "M-Y correlation")
points(tissue.path.stats["X-M"], tissue.path.stats["M-Y"], 
    col = "red", pch = 16)
abline(0,1)

path.min <- min(c(all.perms[[tx]][,"path_coef"], tissue.path.coef), na.rm = TRUE)
path.max <- max(c(all.perms[[tx]][,"path_coef"], tissue.path.coef), na.rm = TRUE)
hist(all.perms[[tx]][,"path_coef"], xlim = c(path.min, path.max), 
    main = "Path Coef. Permutations", xlab = "", breaks = 100)
abline(v = tissue.path.coef, col = "red")

dev.off()

```

```{r null_correlations}
#a reviewer asked about pairwise correlations using randomly sampled
#weights. We test that here.
rnd.weight.cor <- matrix(NA, ncol = 3, nrow = num.perm)
for(tx in 1:max.tx){
    true.cor <- cor(model_scores[[tx]])
    for(p in 1:num.perm){
        sampled.weights <- apply(model_scores[[tx]], 2, function(x) sample(x))
        rnd.cor <- cor(sampled.weights)
        rnd.weight.cor[p,] <- rnd.cor[upper.tri(rnd.cor, diag = FALSE)]
    }
}

pdf(here("Results", "Paper", "Revewer_response_perm.pdf"), width = 4, height = 4)
plot(abs(rnd.weight.cor[,1]), abs(rnd.weight.cor[,3]), xlim = c(0, 0.8), ylim = c(0, 0.8), xlab = "", 
    ylab = "", pch = 16, col = rgb(0.1, 0.1, 0.1, alpha = 0.5))
mtext(expression("G"["C"]-"T"["C"]), side = 1, line = 2.5, cex = 1.5)
mtext(expression("T"["C"]-"P"["C"]), side = 2, line = 2.5, cex = 1.5)

tx = 1
points(all.perms[[tx]][,"X-M"], all.perms[[tx]][,"M-Y"], 
    pch = 16, col = tissue.col)
points(tissue.path.stats["X-M"], tissue.path.stats["M-Y"], 
    col = "red", pch = 16)
dev.off()
```

We also looked at the null distributions of transcript
loadings. It would be interesting to be able to select
transcripts with "significant" loadings in each tissue.
However, the following plots show that most observed 
transcript loadings are well outside the null distribution,
so there is no cutoff that gives us a few transcripts to 
analyze.

I had turned off this code because it takes a long time
and results in large files, but I'm turning it back on
to answer reviewer comments. It can be turned back of 
by setting calc_null to FALSE.


**Note about imputations:** We have found after lots of digging
that the Bayesian method of imputing transcripts with local
and distal components is not working. It is merging the local
and distal signals so that they are highly correlated. We
are now doing the simplest possible imputation, which is the
fitted values from the fit1 mixed model using kinship as a
random effect. This procedure is performed in 3b.Imputation.Rmd. 
To calculate the variance explained by the distal component of 
each transcript, we calculate the variance and heritability of
the residuals of that fit1 model. The variance of the residuals
is the amount of transcript variance that is explained by distal
factors and error. By calculating the heritability of the residuals,
we can estimate the proportion of the residual variance that is
due to distal factors. We are implementing that to replace the Bayesian 
imputations we tried earlier.

```{r prep_pop}
calc_null <- TRUE #set to TRUE to calculate null loading
```

```{r load_herit}
#These imputed transcripts are the fitted values from fit1 
#with the nearest marker and a kinship correction
#created by 3a.Imputation.Rmd
imp.files <- get.files(here("Results", "DO", "Imputed_Transcriptomes"), 
    want = c("Local", "Imputation"), full.names = TRUE)
tissue.imp.gen <- lapply(imp.files, readRDS)
#local.var.exp <- lapply(tissue.imp.gen, function(x) apply(x, 2, var))

#also read in distal components of heritability
#created by 3b.Imputation.Rmd
distal.files <- get.files(here("Results", "DO", "Imputed_Transcriptomes"), 
    want = "Distal", dont.want = "Comparison", full.names = TRUE)
distal.var <- lapply(distal.files, readRDS)

names(distal.var) <- names(tissue.imp.gen) <- tissue.names

#calculate the variance explained by the distal component for each transcript
local.var.exp <- lapply(distal.var, function(x) x[,"local_var"])
dist.var.exp <- lapply(distal.var, function(x) x[,"residual_variance"]*x[,"residual_heritability"])

#trait correlation files
trait.cor.files <- list.files(here("Results", "DO", "Transcriptomes"), 
    pattern = "Maximum_Trait_Correlation", full.names = TRUE) #generated in 2a.Kinship_Expression_Traits.Rmd
trait.cor <- lapply(trait.cor.files, readRDS)
names(trait.cor) <- tissue.names

```

## Latent Variable Correlations {.tabset .tabset-fade .tabset-pills}

The following plots show the partial correlations between 
the pairs of scores for all model scores. With perfect 
mediation of X -> M -> Y, the partial correlation between 
X and Y should be 0. We get pretty close here.

```{r plot_cor, results = "asis", fig.height = 6, fig.width = 6}
#pdf("~/Desktop/test.pdf")
for(tx in 1:length(tissue.names)){
    cat("###", tissue.names[tx], "\n")
    #find the correlations between the scores
    curr_scores <- model_scores[[tx]]
    
    #change from generic names to specific names
    if(complete.mediation){
        colnames(curr_scores) <- c("Genome", "Transcriptome", "Traits")
    }else{
        colnames(curr_scores) <- c("Genome", "Traits", "Transcriptome")
    }
    
    path.coef <- path_coef(curr_scores)

    all.pairs <- pair.matrix(1:ncol(curr_scores))
    adj.var <- apply(all.pairs, 1, function(x) setdiff(1:3, x))
    #pdf("~/Desktop/test.pdf") 
    par(mfrow = c(2,2))
    for(i in 1:nrow(all.pairs)){
        var1 <- all.pairs[i,1]
        var2 <- all.pairs[i,2]
        adj.by <- adj.var[i]
        plot.with.model(adjust(curr_scores[,var1,drop=FALSE], curr_scores[,adj.by,drop=FALSE]), 
            adjust(curr_scores[,var2,drop=FALSE], curr_scores[,adj.by,drop=FALSE]),
            xlab = colnames(curr_scores)[all.pairs[i,1]], 
            ylab = colnames(curr_scores)[all.pairs[i,2]], report = "cor.test",
            col = tissue.cols[tx])
    }
    #test <- check_signs(curr_scores)
    #pairs(test[[1]])

    #pairs(curr_scores, main = paste(tissue.names[tx], "\nPath  Coef:", round(path.coef, 3)))
    cat("\n\n")
}
#dev.off()
```

```{r cor_fig}
if(max.tx == 4){
    pdf(here("Results", "Paper", "Fig4E.pdf"), width = 12, height = 9)
    layout.mat <- matrix(c(0,1:4,5,8,11,14,17,6,9,12,15,18,7,10,13,16,19), 
        ncol = 5, byrow = TRUE)
    layout(layout.mat, widths = c(0.2, rep(1, 4)), 
        heights = c(0.1, rep(1, 3), 0.2))
    #add tissue headers
    for(tx in 1:length(tissue.names)){
        par(mar = c(0,4,0,0))
        plot.text(tissue.names[tx], cex = 2, font = 2, adj = 0.5, x = 0.4, y = 0.25)
    }
    #add plot labels
    gvt <- bquote(italic(G[C])~"-"~italic(T[C]))
    tvp <- bquote(italic(T[C])~"-"~italic(P[C]))
    gvp <- bquote(italic(G[C])~"-"~italic(P[C]))
    plot.labels <- list(gvt, tvp, gvp)
    for(l in 1:length(plot.labels)){
        par(mar = c(4,0,0,0))
        plot.text(plot.labels[[l]], x = 0.3 , y = 0.5, cex = 2, font = 2, srt = 90, adj = 0.5)
    }

    panel.counter <- 8
    for(tx in 1:max.tx){
        if(panel.counter %in% c(8, 11, 14, 17)){
            par(mar = c(4,3,1,3))
        }else{
            par(mar = c(4,3,3,3))
        }
        #find the correlations between the scores
        curr_scores <- model_scores[[tx]]
        
        #change from generic names to specific names
        if(complete.mediation){
            colnames(curr_scores) <- c("Genome", "Transcriptome", "Phenotype")
            axis.labels <- list(bquote(italic(G[C])), 
                bquote(italic(T[C])), 
                bquote(italic(P[C])))
        }else{
            colnames(curr_scores) <- c("Genome", "Phenotype", "Transcriptome")
            axis.labels <- c(bquote(italic(G[C])), 
                bquote(italic(P[C])), 
                bquote(italic(T[C])))
        }
        
        path.coef <- path_coef(curr_scores)

        all.pairs <- pair.matrix(1:ncol(curr_scores))
        all.pairs <- all.pairs[c(1,3,2),]
        adj.var <- apply(all.pairs, 1, function(x) setdiff(1:3, x))
        #pdf("~/Desktop/test.pdf") 
        for(i in 1:nrow(all.pairs)){
            var1 <- all.pairs[i,1]
            var2 <- all.pairs[i,2]
            adj.by <- adj.var[i]
            adj.x <- adjust(curr_scores[,var1,drop=FALSE], curr_scores[,adj.by,drop=FALSE])
            adj.y <- adjust(curr_scores[,var2,drop=FALSE], curr_scores[,adj.by,drop=FALSE])
            model <- lm(adj.y~adj.x)
            f <- summary(model)$fstatistic
            pval <- signif(pf(f[1],f[2],f[3],lower.tail=F), 2)
            p.text <- threshold_p(pval, return.text = TRUE)
            plot(adj.x, adj.y, pch = 16, col = tissue.cols[tx], xlab = "", 
                ylab = "")
            abline(model, col = "#3182bd", lwd = 3)
            plot.text(bquote(italic(R^2)~"="~.(signif(cor(adj.y, adj.x), 2))), rel.x = 0.05, rel.y = 0.95, add = TRUE, adj = 0, cex = 1.2)
            plot.text(p.text, rel.x = 0.05, rel.y = 0.88, add = TRUE, adj = 0, cex = 1.2)
            

            #if(panel.counter %in% c(8, 9, 10)){
                #mtext(axis.labels[var2], side = 2, line = 0.5, adj = 0, cex = 0.8)
                #mtext(axis.labels[var1], side = 1, line = 0.5, adj = 0, cex = 0.8)
                mtext(axis.labels[[var2]], side = 2, line = 2.5)
                mtext(axis.labels[[var1]], side = 1, line = 2.5)
            #}
            #if(panel.counter %in% c(11, 12, 13)){
                #mtext(colnames(curr_scores)[var1], side = 1)
            #}
            plot.dim <- par("usr")
            draw.rectangle(plot.dim[1], plot.dim[2], plot.dim[3], plot.dim[4])
            panel.counter <- panel.counter + 1
        }
        #test <- check_signs(curr_scores)
        #pairs(test[[1]])

        #pairs(curr_scores, main = paste(tissue.names[tx], "\nPath  Coef:", round(path.coef, 3)))
        cat("\n\n")
    }
    dev.off()
}else{
    tx = 1
    pdf(here("Results", "Paper", "Fig4E.pdf"), width = 9, height = 2.8)
    #png(here("Results", "Paper", "Fig4E.png"), width = 9, height = 2.8, units = "in", res = 300)
    layout.mat <- matrix(c(1:6), nrow = 2, byrow = TRUE)
    layout(layout.mat, heights = c(0.1, 1))
    par(bg = NA)
    #add headers for correlation comparisons
    
    gvt <- bquote(italic(G[C])~"-"~italic(T[C]))
    tvp <- bquote(italic(T[C])~"-"~italic(P[C]))
    gvp <- bquote(italic(G[C])~"-"~italic(P[C]))
    plot.labels <- list(gvt, tvp, gvp)
    for(l in 1:length(plot.labels)){
        par(mar = c(0,4,0,4))
        plot.text(plot.labels[[l]], x = 0.5 , y = 0.1, cex = 2, font = 2, adj = c(0.5, 0))
    }

    #find the correlations between the scores
    curr_scores <- model_scores[[tx]]
    
    #change from generic names to specific names
    if(complete.mediation){
        colnames(curr_scores) <- c("Genome", "Transcriptome", "Phenotype")
        axis.labels <- list(bquote(italic(G[C])), 
                bquote(italic(T[C])), 
                bquote(italic(P[C])))
    }else{
        colnames(curr_scores) <- c("Genome", "Phenotype", "Transcriptome")
        axis.labels <- c(bquote(italic(G[C])), 
            bquote(italic(P[C])), 
            bquote(italic(T[C])))
    }
    
    path.coef <- path_coef(curr_scores)

    par(mar = c(4,4,0,4))
    all.pairs <- pair.matrix(1:ncol(curr_scores))
    all.pairs <- all.pairs[c(1,3,2),]
    adj.var <- apply(all.pairs, 1, function(x) setdiff(1:3, x))
    #pdf("~/Desktop/test.pdf") 
    for(i in 1:nrow(all.pairs)){
        var1 <- all.pairs[i,1]
        var2 <- all.pairs[i,2]
        adj.by <- adj.var[i]
        adj.x <- adjust(curr_scores[,var1,drop=FALSE], curr_scores[,adj.by,drop=FALSE])
        adj.y <- adjust(curr_scores[,var2,drop=FALSE], curr_scores[,adj.by,drop=FALSE])
        model <- lm(adj.y~adj.x)
        f <- summary(model)$fstatistic
        pval <- signif(pf(f[1],f[2],f[3],lower.tail=F), 2)
        p.text <- threshold_p(pval, return.text = TRUE)
        R2 <- summary(model)$r.squared
        plot(adj.x, adj.y, pch = 16, xlab = "", ylab = "")
        abline(model, col = "#3182bd", lwd = 3)
        plot.text(bquote(italic(R^2)~"="~.(signif(R2, 2))), rel.x = 0.05, rel.y = 0.95, add = TRUE, adj = 0, cex = 1.2)
        plot.text(p.text, rel.x = 0.05, rel.y = 0.88, add = TRUE, adj = 0, cex = 1.2)
        mtext(axis.labels[[var2]], side = 2, line = 2.5)
        mtext(axis.labels[[var1]], side = 1, line = 2.5)
        plot.dim <- par("usr")
        draw.rectangle(plot.dim[1], plot.dim[2], plot.dim[3], plot.dim[4])
    }
    dev.off()

    fig.data <- cbind(rownames(model_scores[[tx]]), model_scores[[tx]])
    colnames(fig.data) <- c("mouseID", "G_C", "T_C", "P_C")
    write_table_with_header(fig.data, file.path(figure.data.dir, "Fig4E.txt"),
        header = "Data for Fig 4E. Composite genome (G_C), composite 
        transcriptome (T_C), and composite phenome (P_C) scores.")
}
```

# Metabolic Disease Index {.tabset .tabset-fade .tabset-pills}

We are calling the composite phenotype score and Metabolic 
Disease Index, or MDI. This score is highly correlated with
the first principal component of the trait matrix, which is
dominated by body weight and insulin resistance. 

The following plot shows the distribution of MDI across
individuals.

```{r mdi_dist}
for(tx in 1:max.tx){
    if(max.tx == 1){
        cat("## All Tissues\n") 
    }else{
        cat("##", tissue.names[tx], "\n")
    }
    hist(model_scores[[tx]][,"Outcome"], breaks = 50, ylab = "Number of Mice",
        xlab = "", main = "")
}
```

```{r fig4B}
pdf(here("Results", "Paper", "Fig4B.pdf"), width = 5, height = 6)
hist(model_scores[[1]][,"Outcome"], breaks = 50, ylab = "Number of Mice",
        xlab = "", main = "")
dev.off()

fig.data <- cbind("mouseID" = rownames(model_scores[[1]]), 
    "Metabolic Disease Index (MDI)" = model_scores[[1]][,"Outcome"])
write_table_with_header(fig.data, file.path(figure.data.dir, "Fig4B.txt"),
    header = "Data for Fig4B. Metabolic disease index (MDI) for each mouse.")
```

# Transcript Loadings

The following plot shows the distributions of transcript
loadings across the tissues. Adipose tissue has the largest
magnitude loadings in both directions suggesting that transcripts
in the adipose tissue are stronger mediators of genetic effects
on traits than transcripts in the other tissues.

```{r loading_vio}
vioplot(transcript_loadings, col = tissue.cols)
abline(h = 0)
```

```{r mediation_test, eval = FALSE}
bw.herit <- as.numeric(est_herit(scaled.pheno[,"Weight_Final"], Kg))
high.tx <- tail(rownames(transcript_loadings[[1]])[order(transcript_loadings[[1]])], 10)[7]
common.ind <- intersect(rownames(scaled.pheno), rownames(scaled.expr[[1]]))
plot.with.model(scaled.pheno[common.ind,"Weight_Final"], scaled.expr[[1]][common.ind,high.tx])
adj.weight <- adjust(scaled.pheno[,"Weight_Final",drop=FALSE], scaled.expr[[1]][,high.tx,drop=FALSE])
as.numeric(est_herit(adj.weight, Kg))
```

The following figure shows the distribution of the absolute values
of the loadings for each tissue. This plot shows that adipose tissue
has the largest loadings overall, while skeletal muscle has the 
smallest loadings.

```{r loading_abs_vio}
vioplot(lapply(transcript_loadings, abs), col = tissue.cols, ylab = "|Loading|")

#boxplot(transcript_loadings, col = tissue.cols);abline(h = 0)
#boxplot(lapply(transcript_loadings, abs), col = tissue.cols)
#stripchart(lapply(transcript_loadings, function(x) round(abs(x), 4)), 
#    vertical = TRUE, method = "stack", col = tissue.cols, pch = 16, offset = 0.03)
```

```{r overlapping_dist}
#try with overlapping distributions
plot_overlapping_dist <- function(tx.loadings){

    tx.dens <- lapply(tx.loadings, function(x) density(x))
    xmin <- min(sapply(tx.dens, function(x) min(x$x)))
    xmax <- max(sapply(tx.dens, function(x) max(x$x)))
    ymin <- min(sapply(tx.dens, function(x) min(x$y)))
    ymax <- max(sapply(tx.dens, function(x) max(x$y)))

    plot.new()
    plot.window(xlim = c(xmin, xmax), ylim = c(ymin, ymax))
    for(tx in 1:length(tx.dens)){
        plot.poly.xy(poly.top.x = tx.dens[[tx]]$x, 
            poly.top.y = tx.dens[[tx]]$y, 
            poly.bottom.x = tx.dens[[tx]]$x, 
            poly.bottom.y = rep(0, length(tx.dens[[tx]]$y)),
            col = hex2rgb(tissue.cols[tx], alpha = 0.5))
    }
    axis(1);axis(2)
    legend("topright", 
        fill = sapply(tissue.cols, function(x) hex2rgb(x, alpha = 0.5)), 
        legend = names(tissue.cols))
    mtext("Transcript Loading", side = 1, line = 2.5)
    mtext("Density", side = 2, line = 2.5)
}

pdf(here("Results", "Paper", "Supp_Fig_transcript_load_comparison.pdf"), width = 10, height = 5)
par(mfrow = c(1,2))
plot_overlapping_dist(transcript_loadings)
plot_overlapping_dist(lapply(transcript_loadings, abs))
dev.off()
```

The following plots show more detail of these distributions.

```{r loading_dist, fig.height = 8, fig.width = 8}

par(mfrow = c(2,2))
for(tx in 1:length(tissue.names)){
    hist(transcript_loadings[[tx]], breaks = 50, xlab = "Loading",
    main = tissue.names[tx], col = tissue.cols[tx])
    abline(v = 0, col = "red")
}
```

```{r fig4C}
pdf(here("Results", "Paper", "Fig4C.pdf"), width = 5, height = 5)
hist(transcript_loadings[[1]], breaks = 50, , xlab = "Loading",
    main = tissue.names[1], col = tissue.cols[1])
    abline(v = 0, col = "#3182bd", lwd = 2)
dev.off()

fig.data <- cbind("geneID" = rownames(transcript_loadings[[1]]), 
    "Loading" = transcript_loadings[[1]][,1])
write_table_with_header(fig.data, file.path(figure.data.dir, "Fig4C.txt"),
    header = "Data for Figure 4C. Transcript loadings for Adipose transcripts.")
```

## Transcript Loadings and Heritability

We are interested in how local and distal heritability
of transcripts relates to their trait relevance. 

We showed in 3a.Imputation.Rmd that local heritability is
negatively correlated with maximum trait correlation. 

Transcript loading is another measure of trait relevance.
However, transcript loading also has heritability baked
into it, so it should be more positively correlated with 
both local and distal variance explained than simple trait 
correlation.


### Loading vs. Maximum Trait Correlation {.tabset .tabset-fade .tabset-pills}

Maximum trait correlation is our measure of trait 
relevance. Transcript loading is another measure
of trait relevance, thus maximum trait correlation 
and transcript loading should be positively correlated. 
However, transcript loading also has heritability
baked into it, so the importance of transcripts can
be up- or down-graded based on how heritable they 
are.

The following plots show the relationship between
maximum trait correlation and transcript loading.
They are highly correlated, particularly for 
adipose tissue.

```{r loading_cor, fig.height = 5, fig.width = 5, results = "asis"}
for(tx in 1:length(tissue.names)){
    cat("####", tissue.names[tx], "\n")
    common.tx <- intersect(rownames(transcript_loadings[[tx]]), names(trait.cor[[tx]]))
    plot_with_marginal_hist(abs(trait.cor[[tx]][common.tx]),
        abs(transcript_loadings[[tx]][common.tx,1]),
        ylab = "|Loading|", xlab = "|Max. Trait Correlation|", 
        main = tissue.names[tx])
    cat("\n\n")
}
```

### Loading vs. Local Variance Explained {.tabset .tabset-fade .tabset-pills}

The following plots show the correlation between local variance
explained and trait loading. This relationship is much flatter
than the relationship between local variance explained and
trait correlation. This is likely because transcript loading
has heritability incorporated into it, so it makes the 
correlation a little bit more positive.

```{r local_v_loading, fig.width = 5, fig.height = 5, results = "asis"}

for(tx in 1:length(tissue.names)){
    cat("####", tissue.names[tx], "\n")
    common.ts <- Reduce("intersect", list(rownames(transcript_loadings[[tx]]), 
        names(local.var.exp[[tx]])))
    
    plot_with_marginal_hist(abs(transcript_loadings[[tx]][common.ts,1]),
        local.var.exp[[tx]][common.ts], 
        xlab = "|Loading|", ylab = "Local Variance Explained",
        main = tissue.names[tx], report = "cor.test",
        col = tissue.cols[tx])
    cat("\n\n")
}
```

```{r local_v_loading_fig}
pdf(file.path(results.dir, "Trait_Relevance_vs_Local_Var.pdf"), width = 6.5, height = 6.5)
max.load <- max(sapply(transcript_loadings, function(x) max(abs(x))))
max.var.exp <- max(sapply(local.var.exp, function(x) max(x, na.rm = TRUE)))

par(mfrow = c(2,2))
for(tx in 1:length(tissue.names)){
    common.ts <- Reduce("intersect", list(rownames(transcript_loadings[[tx]]), 
        names(local.var.exp[[tx]])))
    
    cor.result <- plot.with.model(abs(transcript_loadings[[tx]][common.ts,1]),
        local.var.exp[[tx]][common.ts], xlim = c(0, max.load),
        ylim = c(0, max.var.exp),
        xlab = "|Loading|", ylab = "Local Variance Explained",
        main = tissue.names[tx], report = "cor.test",
        col = tissue.cols[tx])

    text(x = max.load*0.63, y = max.var.exp*0.95, labels = paste("r =", signif(cor.result[1], 2)), adj = 0)
    text(x = max.load*0.63, y = max.var.exp*0.85, labels = paste("p <", threshold_p(signif(cor.result[2], 2))), adj = 0)
}
dev.off()
```


### Loading vs. Distal Variance Explained {.tabset .tabset-fade .tabset-pills}

Transcript loading is much more positively correlated to distal
heritability. This is also by construction. The transcript score
is constructed to relate distal heritability to trait relevance.

```{r distal_v_loading, fig.width = 5, fig.height = 5, results = "asis"}

#pdf("~/Desktop/test.pdf", width = 5, height = 4)
for(tx in 1:length(tissue.names)){
    cat("####", tissue.names[tx], "\n")
    common.ts <- Reduce("intersect", list(rownames(transcript_loadings[[tx]]), 
        names(dist.var.exp[[tx]])))
    
    plot_with_marginal_hist(abs(transcript_loadings[[tx]][common.ts,1]),
        dist.var.exp[[tx]][common.ts], 
        xlab = "|Loading|", ylab = "Distal Variance Explained",
        main = tissue.names[tx], report = "cor.test", col = tissue.cols[tx])
    cat("\n\n")
}
#dev.off()
```

```{r distal_v_loading_fig}
pdf(file.path(results.dir, "Trait_Relevance_vs_Distal_Var.pdf"), width = 6.5, height = 6.5)
max.load <- max(sapply(transcript_loadings, function(x) max(abs(x))))
max.var.exp <- max(sapply(dist.var.exp, function(x) max(x, na.rm = TRUE)))

par(mfrow = c(2,2))
for(tx in 1:length(tissue.names)){
    common.ts <- Reduce("intersect", list(rownames(transcript_loadings[[tx]]), 
        names(local.var.exp[[tx]])))
    
    cor.result <- plot.with.model(abs(transcript_loadings[[tx]][common.ts,1]),
        dist.var.exp[[tx]][common.ts], xlim = c(0, max.load),
        ylim = c(0, max.var.exp),
        xlab = "|Loading|", ylab = "Distal Variance Explained",
        main = tissue.names[tx], report = "cor.test")

    text(x = max.load*0.63, y = max.var.exp*0.95, labels = paste("r =", signif(cor.result[1], 2)), adj = 0)
    text(x = max.load*0.63, y = max.var.exp*0.85, labels = paste("p =", signif(cor.result[2], 2)), adj = 0)
}
dev.off()
```

### Transcript Loading Null Distributions

```{r calc_null_dist}
if(calc_null){
    null.loading.file <- file.path(results.dir, paste0("Null_Loadings_", perm.mat, "_permuted", ".RDS"))

    if(!file.exists(null.loading.file)){
        null_loadings  <- vector(mode = "list", length = length(transcript_loadings))
        names(null_loadings) <- names(transcript_loadings)

        for(tx in 1:length(transcript_loadings)){
            
            tx.r2 <- apply(tissue.imp.gen[[tx]], 2, var)
            common.transcripts <- intersect(names(tx.r2), rownames(transcript_loadings[[tx]]))

            #calculate null loadings
            if(merge.tissues){
                null_loadings[[tx]] <- sapply(1:ncol(null.mediator.scores[[1]]), 
                    function(x) calc_loadings(null.mediator.scores[[1]][,x,drop=FALSE], 
                    scaled.expr[[tx]]))
                rownames(null_loadings[[tx]]) <- rownames(transcript_loadings[[tx]])
            }else{
                null_loadings[[tx]] <- sapply(1:ncol(null.mediator.scores[[tx]]), 
                    function(x) calc_loadings(null.mediator.scores[[tx]][,x,drop=FALSE], 
                    scaled.expr[[tx]]))
                rownames(null_loadings[[tx]]) <- rownames(transcript_loadings[[tx]])
            }
        }
        saveRDS(null_loadings, null.loading.file)
    }else{
        null_loadings <- readRDS(null.loading.file)
    }
}
```

```{r load_dist, fig.width = 12, fig.height = 3}

if(calc_null){
    all.dens <- c(lapply(transcript_loadings, density), lapply(null_loadings, density))
    xmin <- min(sapply(all.dens, function(x) min(x$x)))
    xmax <- max(sapply(all.dens, function(x) max(x$x)))
    ymax <- max(sapply(all.dens, function(x) max(x$y)))

    #quartz(width = 12, height = 3)
    par(mfrow = c(1,4), mar = c(4,4,4,1))
    for(tx in 1:length(transcript_loadings)){
        num.high <- length(which(transcript_loadings[[tx]] > max(null_loadings[[tx]])))
        num.low <- length(which(transcript_loadings[[tx]] < min(null_loadings[[tx]])))
        plot(density(null_loadings[[tx]]), type = "l", col = "gray", lwd = 3,
            xlim = c(xmin, xmax), ylim = c(0, ymax), xlab = "Transcript Loading",
            main = "", cex.axis = 1.5, cex.lab = 1.3)
        mtext(tissue.names[tx], side = 3, line = 1.5, font = 2)
        points(all.dens[[tx]], type = "l", col = tissue.cols[tx], lwd = 3)
        legend("topleft", col = c(tissue.cols[tx], "gray"), legend = c("observed", "null"), 
            lty = 1, lwd = 3, cex = 1.2)
        plot.dim <- par("usr")
        plot.width = plot.dim[2] - plot.dim[1]
        text(x = (plot.dim[2] - (plot.width*0.4)), y = plot.dim[4]*0.3, 
            labels = paste(num.high, "genes"), adj = 0, cex = 1.5)
        text(x = (plot.dim[2] - (plot.width*0.6)), y = plot.dim[4]*0.3, 
            labels = paste(num.low, "genes"), adj = 1, cex = 1.5)
        #abline(v = c(min(null_loadings[[tx]]), max(null_loadings[[tx]])))
        upper.idx <- which(all.dens[[tx]]$x > max(null_loadings[[tx]]))
        plot.poly.xy(poly.top.x = all.dens[[tx]]$x[upper.idx], 
            poly.top.y = all.dens[[tx]]$y[upper.idx],
            poly.bottom.x = all.dens[[tx]]$x[upper.idx], 
            poly.bottom.y = rep(0, length(upper.idx)), 
            col = hex2rgb(tissue.cols[tx], alpha = 0.5),
            border = NA)

        lower.idx <- which(all.dens[[tx]]$x < min(null_loadings[[tx]]))
        plot.poly.xy(poly.top.x = all.dens[[tx]]$x[lower.idx], 
            poly.top.y = all.dens[[tx]]$y[lower.idx],
            poly.bottom.x = all.dens[[tx]]$x[lower.idx], 
            poly.bottom.y = rep(0, length(lower.idx)), 
            col = hex2rgb(tissue.cols[tx], alpha = 0.5),
            border = NA)

    }
}

```

```{r fig4D}
pdf(here("Results", "Paper", "Fig4D.pdf"), width = 12, height = 3)

if(calc_null){
    all.dens <- c(lapply(transcript_loadings, density), lapply(null_loadings, density))
    xmin <- min(sapply(all.dens, function(x) min(x$x)))
    xmax <- max(sapply(all.dens, function(x) max(x$x)))
    ymax <- max(sapply(all.dens, function(x) max(x$y)))

    #quartz(width = 12, height = 3)
    par(mfrow = c(1,4), mar = c(4,4,4,1))
    for(tx in 1:length(transcript_loadings)){
        num.high <- length(which(transcript_loadings[[tx]] > max(null_loadings[[tx]])))
        num.low <- length(which(transcript_loadings[[tx]] < min(null_loadings[[tx]])))
        plot(density(null_loadings[[tx]]), type = "l", col = "gray", lwd = 3,
            xlim = c(xmin, xmax), ylim = c(0, ymax), xlab = "Transcript Loading",
            main = "", cex.axis = 1.5, cex.lab = 1.3)
        mtext(tissue.names[tx], side = 3, line = 1.5, font = 2)
        points(all.dens[[tx]], type = "l", col = tissue.cols[tx], lwd = 3)
        legend("topleft", col = c(tissue.cols[tx], "gray"), legend = c("observed", "null"), 
            lty = 1, lwd = 3, cex = 1.2)
        plot.dim <- par("usr")
        plot.width = plot.dim[2] - plot.dim[1]
        text(x = (plot.dim[2] - (plot.width*0.4)), y = plot.dim[4]*0.3, 
            labels = paste(num.high, "genes"), adj = 0, cex = 1.5)
        text(x = (plot.dim[2] - (plot.width*0.6)), y = plot.dim[4]*0.3, 
            labels = paste(num.low, "genes"), adj = 1, cex = 1.5)
        #abline(v = c(min(null_loadings[[tx]]), max(null_loadings[[tx]])))
        upper.idx <- which(all.dens[[tx]]$x > max(null_loadings[[tx]]))
        plot.poly.xy(poly.top.x = all.dens[[tx]]$x[upper.idx], 
            poly.top.y = all.dens[[tx]]$y[upper.idx],
            poly.bottom.x = all.dens[[tx]]$x[upper.idx], 
            poly.bottom.y = rep(0, length(upper.idx)), 
            col = hex2rgb(tissue.cols[tx], alpha = 0.5),
            border = NA)

        lower.idx <- which(all.dens[[tx]]$x < min(null_loadings[[tx]]))
        plot.poly.xy(poly.top.x = all.dens[[tx]]$x[lower.idx], 
            poly.top.y = all.dens[[tx]]$y[lower.idx],
            poly.bottom.x = all.dens[[tx]]$x[lower.idx], 
            poly.bottom.y = rep(0, length(lower.idx)), 
            col = hex2rgb(tissue.cols[tx], alpha = 0.5),
            border = NA)

    }
}
dev.off()


tissue.label <- unlist(lapply(1:length(transcript_loadings), function(x) rep(names(transcript_loadings)[x], nrow(transcript_loadings[[x]]))))
tx.label <- unlist(lapply(transcript_loadings, rownames))
fig.data <- cbind("Tissue" = tissue.label, "geneID" = tx.label, "Loading" = unlist(transcript_loadings))
write_table_with_header(fig.data, file.path(figure.data.dir, "Fig4D_observed.txt"),
    header = "Data for Figure 4D. Observed transcript loadings 
    across all tissues. These data are also shown in Supplementary Figure 3.")


null.data <- Reduce("rbind", null_loadings)
colnames(null.data) <- paste0("Perm_", 1:ncol(null.data))
fig.data <- data.frame("Tissue" = tissue.label, "Transcript" = tx.label, null.data)
write_table_with_header(fig.data, file.path(figure.data.dir, "Fig4D_null.txt"),
    header = "Data for Figure 4D. Null transcript loadings across 
    all tissues for 1000 permutations.")
```


## Enrichment of genes with high loadings {.tabset .tabset-fade .tabset-pills}

The following plots show the functional enrichment for 
transcripts at the top and bottom of each list for each 
tissue.

This section also writes out the top and bottom transcripts
for analysis in CMAP.

```{r enrichment}
enrich.file <- file.path(results.dir, paste0("Top_Bottom_Loading_Enrichment_", adjust_numeral, ".RDS"))
hum.mus.ortho <- as.matrix(read.delim(here("Data", "general", "human.mouse.orthologs.txt"), 
    stringsAsFactors = FALSE))

if(!file.exists(enrich.file) || delete_previous){
    all.enrich <- vector(mode = "list", length = length(tissue.names))
    names(all.enrich) <- tissue.names

    top.n <- 200
    for(tx in 1:length(tissue.names)){
        top.genes <- sort(transcript_loadings[[tx]][,1], decreasing = TRUE)[1:top.n]
        top.enrich <- gost(names(top.genes), organism = "mmusculus", source = c("GO", "KEGG", "REACTOME", "HP", "CORUM"))
        
        top.hum.id <- mouse_to_human_ensembl(mouse_ensembl = names(top.genes), 
            hum.mus.ortho = hum.mus.ortho)

        bottom.genes <- sort(transcript_loadings[[tx]][,1], decreasing = FALSE)[1:top.n]
        bottom.enrich <- gost(names(bottom.genes), organism = "mmusculus", source = c("GO", "KEGG", "REACTOME", "HP", "CORUM"))
        bottom.hum.id <- mouse_to_human_ensembl(mouse_ensembl = names(bottom.genes), 
            hum.mus.ortho = hum.mus.ortho)
        
        cmap.table <- cbind(top.hum.id[,"Human.Entrez"], bottom.hum.id[,"Human.Entrez"])
        colnames(cmap.table) <- c("Positive", "Negative")
        write.table(cmap.table, 
            file.path(results.dir, paste0("CMAP_Transcript_Loadings_", tissue.names[tx], "_", adjust_numeral, ".txt")),
            quote = FALSE, row.names = FALSE, sep = "\t")

        if(tx == 1 || tx == 2){
            up.down.trimmed <- lapply(1:2, function(x) cmap.table[which(!is.na(cmap.table[,x]))[1:150],x])
            up.down.label <- unlist(lapply(1:2, function(x) rep(c("up", "down")[x], 150)))
            fig.data <- data.frame("Up_Down" = up.down.label, "Human_entrez_gene_ID" = unlist(up.down.trimmed))
            fig.num <- c("10-12", "11-13")[tx]
            write_table_with_header(fig.data, file.path(figure.data.dir, paste0("Supp_Fig", fig.num, ".txt")),
                header = paste("Data corresponding to Supplementary Figures", 
                gsub("-", " and ", fig.num), 
                "Table includes 150 genes with positive loadings and 150 genes with 
                negative loadings in", tissue.names[tx], "for input into the clue.io
                associated with the connectivity map of the Broad Institute. The clue.io
                tool generated the tables seen in Supplementary Figures", 
                gsub("-", " and ", fig.num)))
        }

        all.enrich[[tx]] <- list("top" = top.enrich, "bottom" = bottom.enrich)
    }
    saveRDS(all.enrich, enrich.file)
}else{
    all.enrich <- readRDS(enrich.file)
}
```


```{r exclude_genes}
#in looking for literature support for each gene,
#there are some genes whose names have other meanings and are related
#to diabetes/obesity. FAU also stands for fatty acid uptake, and CAD
#also stands for coronary artery disease. I did not find any abstracts
#that used these terms as genes, so I am excluding them.
#There are other genes, like Rpl27 that are used as reference genes
#in quantifications of other functional genes. We remove all genes
#with multiple abstracts, but no real literature support.
exclude.genes <- c("Fau", "Cad", "Rpl27")
```

```{r plot_enrichment, results = "asis", fig.width = 12, fig.height = 5, warning = FALSE, message = FALSE}
#pdf("~/Desktop/Enrichment_loading_wordcloud.pdf", width = 10, height = 5)
gene.info <- read.delim(here("Data", "general", "mouse_gene_info.txt"))
lookup.genes = TRUE
search.terms <- c("diabetes", "obesity")
search.location <- "[Title/Abstract]"
plot.top <- 10
lit.thresh = 2
max.term.size = 3000
all.tx.herit <- lapply(tissue.names, 
    function(x) readRDS(here("Results", "DO", "Imputed_Transcriptomes", 
    paste0("Distal_Heritability_", x, ".RDS"))))
ymax <- 0.75
gene.stats <- vector(mode = "list", length = length(tissue.names))
names(gene.stats) <- tissue.names

for(tx in 1:length(tissue.names)){
    tx.herit <- all.tx.herit[[tx]]

    cat("###", tissue.names[tx], "{.tabset .tabset-fade .tabset-pills}\n")

    cat("#### Positive Loadings\n")
    layout(matrix(c(1,2,3,1,2,4), nrow = 2, byrow = TRUE))
    plot.enrichment.wordcloud(enrichment = all.enrich[[tx]]$top, 
        order.by = "p_value",
        plot.label = paste(tissue.names[tx], "Positive Loadings"), 
        num.terms = 25, max.term.size = max.term.size, 
        vertex.col = hex2rgb(tissue.cols[tx], alpha = 0.5))
    
    top.genes <- sort(transcript_loadings[[tx]][,1], decreasing = TRUE)[1:plot.top]
    herit.idx <- match(names(top.genes), rownames(tx.herit))
    local.up.herit <- tx.herit[herit.idx,1]
    distal.up.herit <- tx.herit[herit.idx,3]
    top.name <- gene.tables[[tx]]$symbol[match(names(top.genes), gene.tables[[tx]]$gene.id)]

    up.table <- cbind("Loading" = top.genes, "Local_Heritability" = local.up.herit, 
        "Distal_Heritability" = distal.up.herit)
    rownames(up.table) <- top.name

    par(mar = c(0,4,8,2))
    barplot(rbind(local.up.herit, distal.up.herit), beside = TRUE, 
        las = 2, col = c("gray10", "gray90"), ylab = "",
        names = rep("", plot.top), ylim = c(0, ymax))
    mtext("Heritability", side = 2, line = 3)
    legend("topleft", fill = c("gray10", "gray90"), 
        legend = c("Local", "Distal"))
    
    par(mar = c(8,4,0,2))
    a <- barplot(top.genes*-1, names = rep("", plot.top), las = 2,
        ylim = c(-0.04, 0), axes = FALSE)
    axis(2, at = seq(0, -0.04, -0.01), labels = seq(0, 0.04, 0.01), las = 2)
    mtext("Loading", side = 2, line = 3)
    par(xpd = NA)
    label.y <- max(top.genes)*-1.01
    text(x = a[,1], y = label.y, labels = top.name, adj = 1, srt = 90)
    par(xpd = TRUE)

    if(lookup.genes){
        pubmed.dir <- file.path(results.dir, "PubMed")
        if(!file.exists(pubmed.dir)){dir.create(pubmed.dir, recursive = TRUE)}
        
        tissue.term <- gsub("Skeletal", "", tissue.names[tx])
        gene.term <- paste0(top.name, search.location)
        search.term <- paste0("(", paste(search.terms, collapse = " OR "), ")")
        all.pubmed <- lapply(gene.term, 
            function(x) get_pubmed_ids(paste(c(search.term, tissue.term, x), collapse = " AND ")))

        num.papers <- unlist(lapply(all.pubmed, function(x) x$Count))
        excluded.idx <- which(top.name %in% exclude.genes)
        if(length(excluded.idx) > 0){
            num.papers[excluded.idx] <- 0
        }

        up.table <- cbind(up.table, "Num_Abstracts" = as.numeric(num.papers))

        with.papers <- which(num.papers != 0)
        paper.counts <- cbind(top.name[with.papers], num.papers[with.papers])
        number.order <- order(as.numeric(paper.counts[,2]))

        pdf(file.path(pubmed.dir, paste0("PubMed_Query_", 
            tissue.names[tx], "_Up.pdf")))
        barplot_with_num(as.numeric(paper.counts[number.order,2]), 
            names = paper.counts[number.order,1], las = 2, ylab = "Number of Abstracts")
        dev.off()

        has.data <- all.pubmed[with.papers]
        all.abstracts <- lapply(has.data, function(x) fetch_pubmed_data(x, format = "abstract"))
        for(i in 1:length(all.abstracts)){
            write.table(all.abstracts[[i]], file.path(pubmed.dir, 
                paste0("PubMed_Query_", tissue.names[tx], "_Up_", rownames(up.table)[with.papers[i]], ".txt")), 
                sep = "\t", quote = FALSE, row.names = FALSE, col.names = FALSE)
        }
    }

    cat("\n\n")

    cat("#### Negative Loadings\n")
    layout(matrix(c(1,2,3,1,2,4), nrow = 2, byrow = TRUE))
    plot.enrichment.wordcloud(all.enrich[[tx]]$bottom, 
        plot.label = paste(tissue.names[tx], "Negative Loadings"), 
        num.terms = 25, max.term.size = max.term.size)

    bottom.genes <- sort(transcript_loadings[[tx]][,1], decreasing = FALSE)[1:plot.top]
    herit.idx <- match(names(bottom.genes), rownames(tx.herit))
    local.down.herit <- tx.herit[herit.idx,1]
    distal.down.herit <- tx.herit[herit.idx,3]
    bottom.name <- gene.tables[[tx]]$symbol[match(names(bottom.genes), gene.tables[[tx]]$gene.id)]

    down.table <- cbind("Loading" = bottom.genes, "Local_Heritability" = local.down.herit, 
        "Distal_Heritability" = distal.down.herit)
    rownames(down.table) <- bottom.name

    par(mar = c(0,4,8,2))
    barplot(rbind(local.down.herit, distal.down.herit), beside = TRUE, 
        las = 2, col = c("gray10", "gray90"), ylim = c(0, ymax),
        names = rep("", plot.top))
    mtext("Heritability", side = 2, line = 3)
    legend("topleft", fill = c("gray10", "gray90"), 
        legend = c("Local", "Distal"))
    
    par(mar = c(8,4,0,2))
    a <- barplot(bottom.genes, names = rep("", plot.top), axes = FALSE, ylim = c(-0.04, 0))
    axis(2, at = seq(0, -0.04, -0.02), labels = seq(0, -0.04, -0.02), las = 2)
    mtext("Loading", side = 2, line = 2.5)
    par(xpd = NA)
    label.y <- min(bottom.genes)*1.01
    text(x = a[,1], y = label.y, labels = bottom.name, adj = 1, srt = 90)
    par(xpd = TRUE)

    if(lookup.genes){
        tissue.term <- gsub("Skeletal", "", tissue.names[tx])
        gene.term <- paste0(bottom.name, search.location)
        search.term <- paste0("(", paste(search.terms, collapse = " OR "), ")")
        all.pubmed <- lapply(gene.term, 
            function(x) get_pubmed_ids(paste(c(search.term, tissue.term, x), collapse = " AND ")))

        num.papers <- unlist(lapply(all.pubmed, function(x) x$Count))
        excluded.idx <- which(bottom.name %in% exclude.genes)
        if(length(excluded.idx) > 0){
            num.papers[excluded.idx] <- 0
        }

        down.table <- cbind(down.table, "Num_Abstracts" = as.numeric(num.papers))
        with.papers <- which(num.papers != 0)

        paper.counts <- cbind(bottom.name[with.papers], num.papers[with.papers])
        
        if(length(with.papers) > 0){
            number.order <- order(as.numeric(paper.counts[,2]))
            pdf(file.path(pubmed.dir, paste0("PubMed_Query_", 
                tissue.names[tx], "_Down.pdf")))
            barplot_with_num(as.numeric(paper.counts[number.order,2]), 
                names = paper.counts[number.order,1], las = 2, ylab = "Number of Abstracts")
            dev.off()
        

            has.data <- all.pubmed[with.papers]
            all.abstracts <- lapply(has.data, function(x) fetch_pubmed_data(x, format = "abstract"))
            for(i in 1:length(all.abstracts)){
                write.table(all.abstracts[[i]], file.path(pubmed.dir, 
                    paste0("PubMed_Query_", tissue.names[tx], "_Down_", 
                    rownames(down.table)[with.papers[i]], ".txt")), 
                    sep = "\t", quote = FALSE, row.names = FALSE, col.names = FALSE)
            }
        }
    
        tx.gene.stats <- rbind(up.table, down.table[nrow(down.table):1,])
        gene.stats[[tx]] <- tx.gene.stats
    
    }

    cat("\n\n")
}
#dev.off()
```

The following plots show enrichments across all groups.

```{r group_enrich, fig.width = 12, fig.height = 7}
all.enrich.list <- vector(mode = "list", length = length(tissue.names)*2)
all.enrich.names <- apply(cbind(rep(tissue.names, each = 2), rep(c("Top", "Bottom"), length(tissue.names))), 1, function(x) paste(x, collapse = "_"))
unlist.enrich <- unlist(all.enrich, recursive = FALSE)
names(unlist.enrich) <- all.enrich.names
up.enrich.mat <- plot.enrichment.group(unlist.enrich[c(1,3,5,7)], cluster_cols = FALSE, n.terms = 10,
    max.term.size = max.term.size, cluster_rows = FALSE, plot.results = FALSE, max.char = 50)
down.enrich.mat <- plot.enrichment.group(unlist.enrich[c(2,4,6,8)], cluster_cols = FALSE, n.terms = 10,
    max.term.size = max.term.size, cluster_rows = FALSE, plot.results = FALSE, max.char = 50)

max.enrich <- max(c(max(up.enrich.mat), max(down.enrich.mat)))
par(mfrow = c(1,2), mar = c(4,16,4,2))
imageWithText(up.enrich.mat, use.pheatmap.colors = TRUE, global.color.scale = TRUE,
    global.min = 0, global.max = max.enrich, show.text = FALSE, 
    main = "Most Positive Loadings", row.text.cex = 0.7, row.text.shift = 0.17,
    col.text.cex = 0.7, col.names = tissue.names, col.text.shift = 0.02)
imageWithText(down.enrich.mat, use.pheatmap.colors = TRUE, global.color.scale = TRUE,
    global.min = 0, global.max = max.enrich, show.text = FALSE,
    main = "Most Negative Loadings", row.text.cex = 0.7, row.text.shift = 0.17,
    col.text.cex = 0.7, col.names = tissue.names, col.text.shift = 0.02)

```

```{r enrichment_tables}
pdf(file.path(results.dir, "Enrichment_Tables.pdf"), width = 12, height = 5)
for(tx in 1:length(tissue.names)){
    par(mfrow = c(1,2))
    plot.enrichment(all.enrich[[tx]]$top, 
        plot.label = "Positive Loadings", text.size = 0.7, 
        max.term.size = max.term.size, max.char = 50)
    plot.enrichment(all.enrich[[tx]]$bottom, 
        plot.label = "Negative Loadings", max.char = 50,
        text.size = 0.7, max.term.size = max.term.size)
    mtext(tissue.names[tx], side = 3, line = -2, outer = TRUE)
}
dev.off()
```


```{r gene_info_summary}
all.herit <- unlist(lapply(gene.stats, function(x) x[,2:3]))
min.herit <- min(all.herit)
max.herit <- max(all.herit)
pdf(file.path(results.dir, "gene_stats_by_tissue.pdf"), width = 12, height = 5)
for(i in 1:length(gene.stats)){
    layout(matrix(c(1,2,3), nrow = 1), widths = c(1, 0.5, 1))
    a <- barplot(rev(gene.stats[[i]][,1]), horiz = TRUE, las = 2, main = "Loading")
    segments(x0 = 0, y0 = 0, y1 = max(a))
    imageWithText(gene.stats[[i]][,2:3], use.pheatmap.colors = TRUE, row.text.shift = 0.5,
        main = "", global.color.scale = TRUE, global.min = min.herit, 
        global.max = max.herit, cex = 1, col.names = c("Local", "Distal"), 
        col.text.rotation = 0, col.text.shift = 0.05, col.text.adj = 0.5)
    mtext("Variance Explained", side = 3, cex = 0.8)
    barplot(rev(gene.stats[[i]][,4]), horiz = TRUE, las = 2, main = "Number of Citations")
    mtext(tissue.names[i], side = 3, line = -2, outer = TRUE)
}
dev.off()

#boxplot(gene.stats[[i]][,2:3])
```

```{r plot_gene_stats}

#gene.set.stats is a list with one element per tissue
#where each element is a table containing the gene statistics
#to be plotted. Each gene is in a row, with an external gene
#name as the row name. The columns are the following: Loading,
#Local_Heritability, Distal_Heritability, Num_Abstracts
herit.cols <- as.matrix(read.delim(here("Data", "general", "heritability_colors.txt"), header = FALSE, row.names = 1))[,1]
local.col <- herit.cols["Local"]
distal.col <- herit.cols["Distal"]


plot_gene_stats <- function(gene.set.stats, loading.lim = NULL, herit.max = 1,
    write.pval = FALSE){
    
    all.info <- Reduce("rbind", gene.set.stats)
    tx.label <- rep(tissue.names, each = plot.top*2)

    #png(file.path(results.dir, "gene_stats_combined.png"), width = 5, height = 12, units = "in", res = 300)
    #clust.order <- hclust(dist(all.info[,2:3]))$order
    tissue.col  <- unlist(lapply(1:length(tissue.names), function(x) rep(tissue.cols[x], length(which(tx.label == tissue.names[x])))))
    by.both <- sort.by.then.by(cbind(as.numeric(as.factor(tx.label)), all.info[,"Loading"]), 
        col.type = c("n", "n"), sort.cols = c(2,1), return.order = TRUE)
    clust.order <- c(c(1:length(tx.label))[by.both[,1]])[by.both[,2]]

    #layout.mat <- matrix(c(1,1,1,2,2,2,3,3,3,3,4,3), ncol = 4)
    #layout.mat <- matrix(c(1,1,2,2,3,2), ncol = 3)
    layout.mat <- matrix(c(1,1,5,2,2,4,3,2,4), ncol = 3)
    layout(layout.mat, widths = c(1, 0.4, 0.4), heights = c(0.25, 1, 0.2))
    par(mar = c(4,4,2,0))
    if(is.null(loading.lim)){
        loading.lim <- get_plot_bounds(min(all.info[,"Loading"]), max(all.info[,"Loading"]))
    }
    par(bg = NA)
    a <- barplot(all.info[clust.order,"Loading"],las = 2, xlab = "Loading", 
        col = tissue.col[clust.order], names = NA,
        horiz = TRUE, xlim = loading.lim, cex.lab = 1.5)
    
    mtext("Loading", side = 3, line = -2)
    #legend("topleft", fill = unique(tissue.col), legend = tissue.names, bty = "n")
    segments(y0 = 0, y1 = max(a), x0 = 0, lwd = 2)
    plot.dim <- par("usr")
    legend(x = plot.dim[1], y = plot.dim[4]*0.96, fill = unique(tissue.col), legend = tissue.names, bty = "n")
    #par(mar = c(0,4,0,2))
    #imageWithText(matrix(as.numeric(as.factor(tx.label[clust.order])), nrow = 1), split.at.vals = TRUE,
    #    split.points = seq(0.5, 4.5, 1), col.scale = c("blue", "red", "green", "purple"),
    #    show.text = FALSE, light.dark = "d", row.names = "Tissue")
    par(mar = c(4,6,2,8))
    has.lit.support <- which(all.info[,"Num_Abstracts"] >= lit.thresh)
    lit.support.font <- rep(1, nrow(all.info))
    names(lit.support.font)  <- rownames(all.info)
    lit.support.font[has.lit.support] <- 2
    #herit.bound <- get_plot_bounds(min(all.info[,2:3]), max(all.info[,2:3]), scale.factor = 20)
    herit.bound <- c(0,herit.max)
    imageWithText(all.info[rev(clust.order),2:3], use.pheatmap.colors = TRUE, col.text.shift = 0.01,
        show.text = FALSE, col.names = c("Local", "Distal"), row.text.shift = 0.5, 
        col.text.rotation = 90, col.text.adj = 1, global.color.scale = TRUE, global.min = herit.bound[1],
        global.max = herit.bound[2], row.text.font = lit.support.font[clust.order])
    mtext("Heritability", side = 3, line = -2)
    #par(mar = c(4,0,2,6))
    #a <- barplot(rev(sqrt(all.info[clust.order,4])), xlab = "Square Root\nNum. Publications", horiz = TRUE, 
    #    names = rep("", length(clust.order)))
    par(mar = c(3,4,6,3))
    imageWithTextColorbar(matrix(seq(herit.bound[1], herit.bound[2], 0.1)), x = 7, y = 90,
        use.pheatmap.colors = TRUE, cex = 1,
        global.color.scale = TRUE, global.min = herit.bound[1], global.max = herit.bound[2], las = 2)
    par(mar = c(4,6,0,4))
    if(write.pval){
        test.p <- aov.by.matrix(all.info[,2:3])$"Pr(>F)"[1]
        if(test.p < 2.12e-16){
            ptext <- paste("p <", signif(threshold_p(test.p), 2))
        }else{
            ptext <- paste("p <", signif(threshold_p(test.p), 2))
        }
    }
    boxplot(all.info[,2:3], names = c("L", "D"), ylim = c(0, herit.max), ylab = "Heritability",
        col = c(local.col, distal.col))
    if(write.pval){
        text(x = 2.4, y = 0.9, labels = ptext, adj = 1)
    }
    par(mar = c(4,4,1,2))
    hist(all.info[,1], main = "", xlab = "Loading", ylab = "", breaks = 20, xlim = loading.lim)

    no.dots <- sapply(strsplit(rownames(all.info), ".", fixed = TRUE), function(x) x[1])
    final.info <- data.frame("Tissue" = tx.label, 
        "GeneName" = no.dots, all.info)
    invisible(final.info)
}
```

```{r plot_high_loading}
#High Loading Genes
png(file.path(results.dir, "gene_stats_combined.png"), width = 5, height = 12, units = "in", res = 300)
load.gene.info <- plot_gene_stats(gene.stats)
dev.off()


#png(here("Results", "Paper", "Fig5A.png"), width = 5, height = 12, units = "in", res = 300)
pdf(here("Results", "Paper", "Fig5A.pdf"), width = 5, height = 12)
load.gene.info <- plot_gene_stats(gene.set.stats = gene.stats)
dev.off()

load.sig.test <- t.test(load.gene.info$Local_Heritability, load.gene.info$Distal_Heritability)

write_table_with_header(load.gene.info, file.path(figure.data.dir, "Fig5A.txt"),
    header = "Data associated with Figure 5A. Genes with the top 10 largest
    loadings in each tissue. Table includes the tissue in which each gene is
    expressed, the gene name, the gene's loading in that tissue, its local 
    heritability in that tissue, and its distal heritability in that tissue.
    The final column reports the number of abstracts in PubMed that include
    the gene name, the tissue name, and at least one of the terms diabetes 
    or obesity")

```

Compare these stats to those for genes with high local heritability

```{r high_local_herit_gene_info}
#collect minimal stats for genes with high heritability.
#take twice as many for each tissue, because heritability
#is not signed 
lookup.genes = TRUE
local.gene.stats <- vector(mode = "list", length = length(tissue.names))
names(local.gene.stats) <- tissue.names

for(tx in 1:length(tissue.names)){
    tx.herit <- all.tx.herit[[tx]]
    
    top.genes <- sort(tx.herit[,"local_var"], decreasing = TRUE)[1:(plot.top*2)]
    load.idx <- match(names(top.genes), rownames(transcript_loadings[[tx]]))
    top.load <- transcript_loadings[[tx]][load.idx,1]
    herit.idx <- match(names(top.genes), rownames(tx.herit))
    local.up.herit <- tx.herit[herit.idx,1]
    distal.up.herit <- tx.herit[herit.idx,3]
    top.name <- gene.tables[[tx]]$symbol[match(names(top.genes), gene.tables[[tx]]$gene.id)]

    up.table <- cbind("Loading" = top.load, "Local_Heritability" = local.up.herit, 
        "Distal_Heritability" = distal.up.herit)
    rownames(up.table) <- top.name

    if(lookup.genes){

        tissue.term <- gsub("Skeletal", "", tissue.names[tx])
        gene.term <- paste0(top.name, search.location)
        search.term <- paste0("(", paste(search.terms, collapse = " OR "), ")")
        all.pubmed <- lapply(gene.term, 
            function(x) get_pubmed_ids(paste(c(search.term, tissue.term, x), collapse = " AND ")))

        num.papers <- unlist(lapply(all.pubmed, function(x) x$Count))
        excluded.idx <- which(top.name %in% exclude.genes)
        if(length(excluded.idx) > 0){
            num.papers[excluded.idx] <- 0
        }

        up.table <- cbind(up.table, "Num_Abstracts" = as.numeric(num.papers))

        with.papers <- which(num.papers != 0)
        paper.counts <- cbind(top.name[with.papers], num.papers[with.papers])
        number.order <- order(as.numeric(paper.counts[,2]))

        has.data <- all.pubmed[with.papers]
        all.abstracts <- lapply(has.data, function(x) fetch_pubmed_data(x, format = "abstract"))
        for(i in 1:length(all.abstracts)){
            write.table(all.abstracts[[i]], file.path(pubmed.dir, 
                paste0("PubMed_Query_", tissue.names[tx], "_High_Local_Herit_", 
                rownames(up.table)[with.papers[i]], ".txt")), 
                sep = "\t", quote = FALSE, row.names = FALSE, col.names = FALSE)
        }
    }
    local.gene.stats[[tx]] <- up.table
}



#Genes with high local heritability
png(file.path(results.dir, "gene_stats_combined_local_herit.png"), width = 5, height = 12, units = "in", res = 300)
local.gene.info <- plot_gene_stats(local.gene.stats, loading.lim = c(-0.06, 0.06))
dev.off()

#png(here("Results", "Paper", "Fig5C.png"), width = 5, height = 12, units = "in", res = 300)
pdf(here("Results", "Paper", "Fig5C.pdf"), width = 5, height = 12)
local.gene.info <- plot_gene_stats(local.gene.stats, loading.lim = c(-0.06, 0.06))
dev.off()

local.sig.test <- t.test(local.gene.info$Local_Heritability, local.gene.info$Distal_Heritability)


write_table_with_header(local.gene.info, file.path(figure.data.dir, "Fig5C.txt"),
    header = "Data associated with Figure 5C. Genes with the top 10 largest
    local heritabilities in each tissue. Table includes the tissue in which 
    each gene is expressed, the gene name, the gene's loading in that tissue, 
    its local heritability in that tissue, and its distal heritability in that 
    tissue. The final column reports the number of abstracts in PubMed that 
    include the gene name, the tissue name, and at least one of the terms 
    diabetes or obesity")


prop.high.loading.lit <- length(which(load.gene.info[,"Num_Abstracts"] >= lit.thresh))/nrow(load.gene.info)
prop.high.herit.lit <- length(which(local.gene.info[,"Num_Abstracts"] >= lit.thresh))/nrow(local.gene.info)

#factor lit increase for high loading
loading.herit.enrich <- prop.high.loading.lit/prop.high.herit.lit

```

Also look at TWAS genes with high correlations to body weight.
This more closely hews to how people select genes, rather than 
high local heritability alone.

```{r high_twas}
#collect stats for genes with significant TWAS
#correlations.

twas.results <- read.csv(here("Results", "DO", "Imputed_Transcriptomes", "TWAS_results.csv"))

lookup.genes = TRUE
twas.gene.stats <- vector(mode = "list", length = length(tissue.names))
names(twas.gene.stats) <- tissue.names

for(tx in 1:length(tissue.names)){
    tx.idx <- which(twas.results[,"Tissue"] == tissue.names[tx])
    tx.cor <- twas.results[tx.idx,"Correlation"]
    names(tx.cor) <- twas.results[tx.idx,"gene.id"]

    #remove duplicates. Some transcripts are associated with more than one trait
    tx.cor <- tx.cor[which(!duplicated(names(tx.cor)))]
    top.genes <- sort(tx.cor, decreasing = TRUE)[1:plot.top]
    
    load.idx <- match(names(top.genes), rownames(transcript_loadings[[tx]]))
    top.load <- transcript_loadings[[tx]][load.idx,1]
    tx.herit <- all.tx.herit[[tx]]
    herit.idx <- match(names(top.genes), rownames(tx.herit))
    local.up.herit <- tx.herit[herit.idx,1]
    distal.up.herit <- tx.herit[herit.idx,3]
    top.name <- gene.tables[[tx]]$symbol[match(names(top.genes), gene.tables[[tx]]$gene.id)]

    up.table <- cbind("Loading" = top.load, "Local_Heritability" = local.up.herit, 
        "Distal_Heritability" = distal.up.herit)
    rownames(up.table) <- top.name

    if(lookup.genes){

        tissue.term <- gsub("Skeletal", "", tissue.names[tx])
        gene.term <- paste0(top.name, search.location)
        search.term <- paste0("(", paste(search.terms, collapse = " OR "), ")")
        all.pubmed <- lapply(gene.term, 
            function(x) get_pubmed_ids(paste(c(search.term, tissue.term, x), collapse = " AND ")))

        num.papers <- unlist(lapply(all.pubmed, function(x) x$Count))
        excluded.idx <- which(top.name %in% exclude.genes)
        if(length(excluded.idx) > 0){
            num.papers[excluded.idx] <- 0
        }

        up.table <- cbind(up.table, "Num_Abstracts" = as.numeric(num.papers))

        with.papers <- which(num.papers != 0)
        
        if(length(with.papers) > 0){
            paper.counts <- cbind(top.name[with.papers], num.papers[with.papers])
            number.order <- order(as.numeric(paper.counts[,2]))
            has.data <- all.pubmed[with.papers]
            all.abstracts <- lapply(has.data, function(x) fetch_pubmed_data(x, format = "abstract"))
            for(i in 1:length(all.abstracts)){
                write.table(all.abstracts[[i]], file.path(pubmed.dir, 
                    paste0("PubMed_Query_", tissue.names[tx], "_TWAS_Positive_Cor_", 
                    rownames(up.table)[with.papers[i]], ".txt")), 
                    sep = "\t", quote = FALSE, row.names = FALSE, col.names = FALSE)
            }
        }
    }

    #look at negatively correlated genes too.
    bottom.genes <- sort(tx.cor, decreasing = FALSE)[1:plot.top]
    
    load.idx <- match(names(bottom.genes), rownames(transcript_loadings[[tx]]))
    bottom.load <- transcript_loadings[[tx]][load.idx,1]
    tx.herit <- all.tx.herit[[tx]]
    herit.idx <- match(names(bottom.genes), rownames(tx.herit))
    local.down.herit <- tx.herit[herit.idx,1]
    distal.down.herit <- tx.herit[herit.idx,3]
    bottom.name <- gene.tables[[tx]]$symbol[match(names(bottom.genes), gene.tables[[tx]]$gene.id)]

    down.table <- cbind("Loading" = bottom.load, "Local_Heritability" = local.down.herit, 
        "Distal_Heritability" = distal.down.herit)
    rownames(down.table) <- bottom.name

    if(lookup.genes){

        tissue.term <- gsub("Skeletal", "", tissue.names[tx])
        gene.term <- paste0(bottom.name, search.location)
        search.term <- paste0("(", paste(search.terms, collapse = " OR "), ")")
        all.pubmed <- lapply(gene.term, 
            function(x) get_pubmed_ids(paste(c(search.term, tissue.term, x), collapse = " AND ")))

        num.papers <- unlist(lapply(all.pubmed, function(x) x$Count))
        excluded.idx <- which(bottom.name %in% exclude.genes)
        if(length(excluded.idx) > 0){
            num.papers[excluded.idx] <- 0
        }

        down.table <- cbind(down.table, "Num_Abstracts" = as.numeric(num.papers))

        with.papers <- which(num.papers != 0)
        
        if(length(with.papers) > 0){
            paper.counts <- cbind(bottom.name[with.papers], num.papers[with.papers])
            number.order <- order(as.numeric(paper.counts[,2]))
            has.data <- all.pubmed[with.papers]
            all.abstracts <- lapply(has.data, function(x) fetch_pubmed_data(x, format = "abstract"))
            for(i in 1:length(all.abstracts)){
                write.table(all.abstracts[[i]], file.path(pubmed.dir, 
                    paste0("PubMed_Query_", tissue.names[tx], "_TWAS_Negative_Cor_", 
                    rownames(down.table)[with.papers[i]], ".txt")), 
                    sep = "\t", quote = FALSE, row.names = FALSE, col.names = FALSE)
            }
        }
    }

    twas.gene.stats[[tx]] <- rbind(up.table, down.table)
}


#Genes with high local heritability
png(file.path(results.dir, "gene_stats_combined_twas.png"), width = 5, height = 12, units = "in", res = 300)
twas.gene.info <- plot_gene_stats(twas.gene.stats, loading.lim = c(-0.06, 0.06))
dev.off()

#png(here("Results", "Paper", "Fig5B.png"), width = 5, height = 12, units = "in", res = 300)
pdf(here("Results", "Paper", "Fig5B.pdf"), width = 5, height = 12)
twas.gene.info <- plot_gene_stats(twas.gene.stats, loading.lim = c(-0.06, 0.06))
dev.off()

twas.sig.test <- t.test(twas.gene.info$Local_Heritability, twas.gene.info$Distal_Heritability)

write_table_with_header(twas.gene.info, file.path(figure.data.dir, "Fig5B.txt"),
    header = "Data associated with Figure 5B. Genes with the top 10 largest
    TWAS scores in each tissue. Table includes the tissue in which each gene 
    is expressed, the gene name, the gene's loading in that tissue, its local 
    heritability in that tissue, and its distal heritability in that tissue.
    The final column reports the number of abstracts in PubMed that include
    the gene name, the tissue name, and at least one of the terms diabetes 
    or obesity")

prop.high.loading.lit <- length(which(load.gene.info[,"Num_Abstracts"] >= lit.thresh))/nrow(load.gene.info)
prop.high.twas.lit <- length(which(twas.gene.info[,"Num_Abstracts"] >= lit.thresh))/nrow(twas.gene.info)

#factor lit increase for high loading
loading.twas.enrich <- prop.high.loading.lit/prop.high.twas.lit

saveRDS(c("loading.twas.enrich" = loading.twas.enrich, "loading.herit.enrich" = loading.herit.enrich),
    here("Results", "Paper", "Lit_enrich.RDS"))

```


```{r enrich_tables, eval = FALSE}
pdf("~/Desktop/Loading_Enrichment.pdf", width = 20, height = 9)
for(tx in 1:length(tissue.names)){

    par(mfrow = c(1,2))
    plot.enrichment(enrichment = all.enrich[[tx]]$top, 
        order.by = "p_value",
        plot.label = paste(tissue.names[tx], "Positive Loadings"), 
        num.terms = 25, max.term.size = 3000)

    plot.enrichment(all.enrich[[tx]]$bottom, 
        plot.label = paste(tissue.names[tx], "Negative Loadings"), 
        num.terms = 25, max.term.size = 3000)

    write.table(all.enrich[[tx]]$top$result[,c("term_name", "term_id", "p_value")],
        "~/Desktop/top_enrich.txt", sep = "\t", row.names = FALSE,
        quote = FALSE)

    write.table(all.enrich[[tx]]$bottom$result[,c("term_name", "term_id", "p_value")],
        "~/Desktop/bottom_enrich.txt", sep = "\t", row.names = FALSE, 
        quote = FALSE)

}
dev.off()

```

# Trait loadings

The following plot shows the loadings on traits.
This shows which traits are most affected by the 
genome -> transcriptome path.

```{r trait_loadings, fig.width = 3.5, fig.height = 6}
all.loadings <- Reduce("cbind", trait_loadings)
colnames(all.loadings) <- tissue.names

#pdf("~/Desktop/test.pdf", width = 7, height = 7)
if(merge.tissues){
    par(mar = c(4, 8, 4, 2))
    barplot(sort(all.loadings[,1]), las = 2, horiz = TRUE, xlab = "Trait Loading")
}else{
    pheatmap(all.loadings)
}
#dev.off()
```

```{r fig4A}
pdf(here("Results", "Paper", "Fig4A.pdf"), width = 3.5, height = 6)
par(mar = c(4, 8, 4, 2))
barplot(sort(all.loadings[,1]), las = 2, horiz = TRUE, xlab = "Trait Loading")
dev.off()

fig.table <- cbind("Trait" = rownames(all.loadings), "Loading" = all.loadings[,1])
write_table_with_header(fig.table, file.path(figure.data.dir, "Fig4A.txt"),
    header = "Data for Figure 4A. Trait loadings.")
```

# Polytranscriptomic risk scores

Can we use each animal's transcription profile to predict 
their phenotype? Here, we multiply the transcript loadings
by each individual's transcription profile and calculate the
mean as its predicted position in phenotype space.

We do this for both the measured transcripts and the imputed
transcripts. The imputed transcripts carry only local genetic
information.

## Measured Transcripts

The following plots show the correlations between phenome score
and the predicted phenome score based on individual measured 
transcriptome.

```{r ptrs, fig.width = 8, fig.height = 8}
if(mediation.text == "complete_mediation"){
    transcript.pred <- vector(mode = "list", length = length(tissue.names))
    names(transcript.pred) <- tissue.names

    par(mfrow = c(2,2))
    for(tx in 1:length(tissue.names)){

        tx.idx <- which(names(tissue.imp.gen) == tissue.names[tx])
        if(length(tx.idx) == 0){next()} #we don't have all the tissues imputed yet.

        #calculate a loaded transcript score for every individual
        #this is their measured transcriptome times the loadings.
        #first scale expression across individuals
        ind.scaled.expr <- apply(adj.expr[[tx]], 2, scale)
        dimnames(ind.scaled.expr) <- dimnames(adj.expr[[tx]])
        
        common.tx <- intersect(colnames(ind.scaled.expr), rownames(transcript_loadings[[tx]]))
        transcript.pred[[tx]] <- apply(ind.scaled.expr[,common.tx], 1, function(x) x * transcript_loadings[[tx]][common.tx,])
        mean.pred <- colMeans(transcript.pred[[tx]])
        pred.order <- order(mean.pred)    
        
        common.ind <- intersect(rownames(ind.scaled.expr), rownames(model_scores[[tx]]))
        plot.with.model(mean.pred[common.ind], model_scores[[tx]][,"Outcome"],
            xlab = "Prediction", ylab = "Phenome Score", 
            main = tissue.names[tx], col = tissue.cols[tx])
    }

}
```

## Imputed Transcripts

The following plots show the correlations between phenome score
and the predicted phenome score based on individual imputed 
transcriptome.

```{r imputed_prediction, fig.width = 8, fig.height = 8}
if(mediation.text == "complete_mediation"){
    transcript.pred <- vector(mode = "list", length = length(tissue.names))
    names(transcript.pred) <- tissue.names

    par(mfrow = c(2,2))
    for(tx in 1:length(tissue.names)){

        tx.idx <- which(names(tissue.imp.gen) == tissue.names[tx])
        if(length(tx.idx) == 0){next()} #we don't have all the tissues imputed yet.
        
        common.tx <- intersect(colnames(tissue.imp.gen[[tx.idx]]), rownames(transcript_loadings[[tx]]))
        transcript.pred[[tx]] <- apply(tissue.imp.gen[[tx.idx]][,common.tx], 1, function(x) x * transcript_loadings[[tx]][common.tx,])
        mean.pred <- colMeans(transcript.pred[[tx]], na.rm = TRUE)
        pred.order <- order(mean.pred)    
        
        common.ind <- intersect(rownames(ind.scaled.expr), rownames(model_scores[[tx]]))
        plot.with.model(mean.pred[common.ind], 
            model_scores[[tx]][common.ind,"Outcome"],
            xlab = "Prediction", ylab = "Phenome Score", 
            main = tissue.names[tx], col = tissue.cols[tx])
    }
}
```

## Distributions for individual extremes  {.tabset .tabset-fade .tabset-pills}

The following histograms show the transcriptional profiles of the 
fattest and leanest individuals. Values shown are the loaded transcripts
for an individual indicating whether each transcript is increasing or 
decreasing the individuals obesity. Below the histograms are boxplots 
showing these distributions for the top and bottom individuals based 
on phenome score.


We also plot to pngs matrices showing the proportion of 
obesity-exacerbating (red), and obesity-alleviating (blue) 
transcripts based on their loadings (Risk_Count_tissue_name.png). 
The individuals are sorted by their phenotype score. There are 
proportionally more exacerbating transcripts in the animals with
higher phenotype scores than in those with lower scores.


```{r polygenic_risk_score, warning = FALSE, message = FALSE, fig.width = 8, fig.height = 8, results = "asis"}
if(mediation.text == "complete_mediation"){
    pred.mats <- vector(mode = "list", length = length(tissue.names))
    names(pred.mats) <- tissue.names

    for(tx in 1:length(tissue.names)){
        cat("###", tissue.names[tx], "\n")
        tx.idx <- which(names(tissue.imp.gen) == tissue.names[tx])
        if(length(tx.idx) == 0){next()} #we don't have all the tissues imputed yet.

        #look at the distribution of loaded transcripts for an individual
        #this is for the individual with the lowest phenome score
        #pdf(file.path(results.dir, paste0("Individual_Comparisons_", tissue.names[tx], "_", 
        #    adjust_numeral, ".pdf")), width = 10, height = 10)
        layout(matrix(c(1,2,3,3), nrow = 2, byrow = TRUE))
        phenome.score <- model_scores[[tx]][,"Outcome"]
        sorted.phenome.score <- sort(phenome.score, decreasing = FALSE)
        
        min.ind <- which.min(phenome.score)
        hist(transcript.pred[[tx]][,names(min.ind)], breaks = 100, xlab = "Loaded Transcript",
            main = "Distribution of Loaded Transcripts\nFor Leanest Animal")
        abline(v = mean(transcript.pred[[tx]][,names(min.ind)]), col = "red")

        max.ind <- which.max(phenome.score)
        hist(transcript.pred[[tx]][,names(max.ind)], breaks = 100, xlab = "Loaded Transcript",
            main = "Distribution of Loaded Transcripts\nFor Fattest Animal")
        abline(v = mean(transcript.pred[[tx]][,names(max.ind)]), col = "red")

        leanest <- names(head(sorted.phenome.score, 10))
        fattest <- names(tail(sorted.phenome.score, 10))
        boxplot(cbind(transcript.pred[[tx]][,c(leanest, fattest)]), las = 2, ylab = "Loaded Transcript",
            main = "Comparison of Top and Bottom 10 of Phenotype Distribution")
        abline(h = 0);abline(v = 10.5)
        #dev.off()

        #boxplot(transcript.pred[,common.ind[pheno.order]])
        pred.mat <- matrix(0, nrow = nrow(transcript.pred[[tx]]), ncol = length(phenome.score))
        colnames(pred.mat) <- names(sorted.phenome.score)
        
        pred.mat[which(transcript.pred[[tx]][,names(sorted.phenome.score)] > 0)] <- 1
        sorted.pred <- apply(pred.mat, 2, sort)
        pred.mats[[tx]] <- sorted.pred
        cat("\n\n")
    }
}
graphics.off() #make sure all graphics devices are shut off
```

```{r plot_prediction_matrices}
for(tx in 1:length(tissue.names)){
    png(file.path(results.dir, paste0("Risk_Count_", tissue.names[tx], "_", 
        adjust_numeral, ".png")), width = 10, height = 5, units = "in", res = 100)
        pheatmap(sorted.pred, cluster_rows = FALSE, cluster_cols = FALSE, show_rownames = FALSE,
            show_colnames = FALSE)
    dev.off()
}


```

# Tissues of Action

How are transcript loadings distributed across tissues?
High loadings in one tissue tells us the tissue of action 
for individual genes. Are there commonalities across tissues?

The following code generates a clustered heatmap of transcript
loadings across tissues for which each transcript has a z score
of at least 2.5 in at least one of the tissues. It looks for 
functional enrichments in each of the clusters and plots the
results both with quick labels for each cluster, as well as 
more detailed information about the enrichments of each cluster.

The plots are printed to a pdf with the title TOA_Transcript_Loadings.pdf

```{r get_max_enrich}
#This function gets the most significant term out of an
#enrichment table returned from gost.

get_max_enrich <- function(enrichment, max.term.size = NULL){
    if(is.null(enrichment)){
        return("no enrichment")
    }

    if(!is.null(max.term.size)){
        below.max <- which(enrichment[[1]][,"term_size"] <= max.term.size)
        if(length(below.max) == 0){return("no enrichment")}
        enrichment[[1]] <- enrichment[[1]][below.max,]
    }

    min.p.idx <- which.min(enrichment[[1]][,"p_value"])
    sig.term <- enrichment[[1]][min.p.idx,"term_name"]
    return(sig.term)
}

#This function takes in a list of loadings.
#The list should consist of named vectors, where the names
#are the transcript names.
#filters them for high values based on the given threshold,
#and clusters into groups. It picks the number of clusters
#that give the best cluster separation in k-means clustering
#It looks for functional enrichments of each cluster, and then
#plots a heatmap of the clusters along with more detailed 
#reports of the enrichment of each.
#z.thresh is the z.score threshold to use for selecting
#transcripts.
cluster_values <- function(value.list, z.thresh = 3, k.seq = 3:20){

    #threshold values above Z score threshold
    scaled.values <- lapply(value.list, rankZ)
    high.values <- lapply(scaled.values, function(x) which(abs(x) > z.thresh))
    all.high <- Reduce("union", lapply(high.values, names))
    all.high.idx <- sapply(value.list, function(x) match(all.high, names(x)))

    #combine into a matrix that contains all unique transcripts
    #above the threshold
    value.mat <- sapply(1:length(value.list), 
        function(x) value.list[[x]][all.high.idx[,x]])
    rownames(value.mat) <- all.high
    colnames(value.mat) <- tissue.names
    
    #remove transcripts not expresed in all tissues
    #keep.vals <- which(!is.na(rowSums(value.mat)))
    #value.mat <- value.mat[keep.vals,]

    #set NAs to 0
    value.mat[which(is.na(value.mat))] <- 0

    #use k-means clustering to figure out how many clusters
    #there are in this matrix. Set the number of clusters to
    #test with k.seq
    #test.mat <- plot.decomp(value.mat, pc = length(tissue.names), plot.results = FALSE)$u
    test.k <- test.pam.k(value.mat, plot.results = FALSE, kseq = k.seq)
    mean.cl.width <- sapply(test.k$cl.width, mean)
    #barplot(mean.cl.width)
    #select the number of clusters with the best separation among clusters.
    k = as.numeric(names(mean.cl.width)[which.max(mean.cl.width)])
    final.clust <- pam(value.mat, k = k)
    #pheatmap(value.mat[names(final.clust$clustering)[order(final.clust$clustering)],], 
    #    cluster_rows = FALSE, cluster_cols = FALSE)
    
    gene.groups <- final.clust$clustering

    group.vals <- lapply(1:k, function(x) value.mat[names(which(gene.groups == x)),,drop=FALSE]) 

    #look for functional enrichment of the gene clusters
    group.gene.names <- lapply(1:k, 
        function(x) names(gene.groups)[which(gene.groups == x)])
    group.enrich <- lapply(group.gene.names, 
        function(x) gost(x, organism = "mmusculus", 
        source = c("GO", "REACTOME", "KEGG", "WP", "CORUM", "HP")))
        
    #select the term with the smallest p value to represent 
    #each group
    group.labels <- sapply(group.enrich, function(x) get_max_enrich(x, max.term.size = 3000))
    names(group.enrich) <- names(group.gene.names) <- group.labels

    results <- list("group_enrich" = group.enrich, "group_genes" = group.gene.names, 
        "loadings" = value.mat[names(final.clust$clustering)[order(final.clust$clustering)],], 
        "grouped_values" = group.vals)
    invisible(results)
}


plot_clustered_values <- function(cluster_results, max.term.size = 3000, 
    enrichment.tables = TRUE, loading.viols = TRUE, alt.labels = NULL,
    cluster.borders = c("boxes", "bars", "both")){

    cluster.borders <- cluster.borders[1]

    group.enrich <- cluster_results$group_enrich
    value.mat <- cluster_results$loadings
    group.gene.names <- cluster_results$group_genes
    
    group.labels <- alt.labels
    if(is.null(group.labels)){
        group.labels <- names(group.enrich)
    }
    group.vals <- cluster_results$grouped_values

    enrich.labels <- lapply(1:length(group.enrich), 
        function(x) rep(group.labels[x], length(group.gene.names[[x]])))
    enrich.label.v <- unlist(enrich.labels)

    u_groups <- unique(unlist(enrich.labels))
    names(u_groups) <- paste("Group", 1:length(u_groups))

    #place cluster labels in the middle of each cluster
    row.labels <- rep("", nrow(value.mat))
    for(g in 1:length(u_groups)){
        g.idx <- which(enrich.label.v == u_groups[g])
        row.labels[median(g.idx)] <- u_groups[g]
    }

    par(mar = c(6,20,2,2))
    layout.mat <- matrix(c(1,2,1,0), nrow = 2, byrow = TRUE)
    layout(layout.mat, widths = c(1,0.2))
    if(cluster.borders == "bars" || cluster.borders == "both"){
        imageWithText(value.mat, show.text = FALSE, use.pheatmap.colors = TRUE, 
            row.names = row.labels, row.text.shift = 0.36, col.text.shift = 0.01)
    }
    
    if(cluster.borders == "boxes"){
        imageWithText(value.mat, show.text = FALSE, use.pheatmap.colors = TRUE, 
            row.names = row.labels, row.text.shift = 0.2, col.text.shift = 0.01)
    }

    plot.dim <- par("usr")

    #draw boxes to delineate clusters
    plot.coord <- nrow(value.mat):1
    for(g in 1:length(u_groups)){
        g.idx <- which(enrich.label.v == u_groups[g])
        if(g %% 2 == 0){
            box.col = "lightgray"
        }else{
            box.col = "white"
        }

        if(cluster.borders == "bars" || cluster.borders == "both"){
            draw.rectangle(0, 0.5, plot.coord[min(g.idx)], plot.coord[max(g.idx)], 
                fill = box.col, border.col = "gray35")
        }
    
        if(cluster.borders == "boxes" || cluster.borders == "both"){
            draw.rectangle(plot.dim[1], plot.dim[2], plot.coord[min(g.idx)], plot.coord[max(g.idx)], 
                fill = NA, border.col = "black")
        }

    }

    #plot a color bar for the heatmap
    par(mar = c(4,3,4,3))
    imageWithTextColorbar(value.mat, use.pheatmap.colors = TRUE, 
        bar.lwd = 3, cex = 1, round.nearest = max(abs(value.mat)/100),
        axis.line = -4)

    if(enrichment.tables){
        #further descriptions of enrichments for groups
        par(mfrow = c(3,1), mar = c(2,4,2,4))
        for(g in 1:length(group.enrich)){
            #group.enrich.name = get_max_enrich(group.enrich[[g]], max.term.size = 3000)
            group.enrich.name <- names(group.enrich)[g]
            plot.enrichment(group.enrich[[g]], 
                plot.label = paste("Group:", group.enrich.name),
                num.terms = 20, order.by = "p_value", max.term.size = max.term.size)
        }

    }

    if(loading.viols){
        #show value distributions per group
        par(mfrow = c(3,2), mar = c(4,4,4,4))
        for(g in 1:length(group.vals)){
            vioplot(group.vals[[g]], main = get_max_enrich(group.enrich[[g]], 
                max.term.size = max.term.size), col = tissue.cols, las = 2)
            abline(h = 0)
        }
    }
    result <- list("value.mat" = value.mat, "labels" = enrich.label.v)
    invisible(result)
}
```

There are lots of very strong loadings among the transcripts. 
These are transcripts that are highly related to the traits,
but do not necessarily have local eQTL in the population.

Transcripts without local eQTL are those that are highly
constrained in this population and not allowed to vary 
based on genetics. But they can still have a large effect
on the trait when they do vary. We would expect genes in this
list to have a large impact if manipulated artificially, say
with a KO.

The loading patterns are quite tissue-specific. Clusters
of genes tend to be active in individual tissues. There are some
shared clusters:
    Adipose and liver share negative loadings in the macromolecule
        biosynthesis cluster
    Adipose and skeletal muscle share positive loadings in the 
        transmembrane activity cluster

```{r toa_load, warning = FALSE, message = FALSE}

transcript.loading.list <- lapply(transcript_loadings, function(x) x[,1])

toa_loadings_file <- file.path(results.dir, paste0("TOA_Transcript_Loadings_", adjust_numeral, "_Results.RDS"))
if(!file.exists(toa_loadings_file)){
   toa.results <- cluster_values(value.list = transcript.loading.list, z.thresh = 2.5, 
        k.seq = 3:20)
    saveRDS(toa.results, toa_loadings_file)
}else{
    toa.results <- readRDS(toa_loadings_file)
}

graphics.off()
pdf(file.path(results.dir, paste0("TOA_Transcript_Loadings_", adjust_numeral, ".pdf")))
val.mat <- plot_clustered_values(toa.results)
dev.off()

#using the enrichment tables plotted in the above pdf, I saw
#that the top enrichments aren't always the best description of the
#cluster. I generated cluster_descriptions.txt by hand as a way to
#create better descriptions. This file is in Data/general, and we
#can use it to annotate the clusters for the figure below.
alt.labels <- read.table(here("Data", "general", "cluster_descriptions.txt"), header = FALSE, sep = "\t")

png(file.path(results.dir, "TOA_Transcript_Loadings_Figure_boxes.png"), width = 6, height = 7, units = "in", res = 300)
plot_clustered_values(toa.results, enrichment.tables = FALSE, loading.viols = FALSE,
    alt.labels = alt.labels[,1], cluster.borders = "boxes")
dev.off()

png(file.path(results.dir, "TOA_Transcript_Loadings_Figure_bars.png"), width = 6, height = 7, units = "in", res = 300)
val.mat <- plot_clustered_values(toa.results, enrichment.tables = FALSE, loading.viols = FALSE,
    alt.labels = alt.labels[,1], cluster.borders = "bars")
dev.off()

png(file.path(results.dir, "TOA_Transcript_Loadings_Figure_both.png"), width = 6, height = 7, units = "in", res = 300)
val.mat <- plot_clustered_values(toa.results, enrichment.tables = FALSE, loading.viols = FALSE,
    alt.labels = alt.labels[,1], cluster.borders = "both")
dev.off()

#write out gene lists for each cluster
toa.cluster.dir <- file.path(results.dir, "TOA_Clusters")
if(!file.exists(toa.cluster.dir)){dir.create(toa.cluster.dir)}
gene.lists <- toa.results[[2]]
if(nrow(alt.labels) == length(gene.lists)){
    names(gene.lists) <- alt.labels[,1]
}
full.list <- vector(mode = "list", length = length(gene.lists))
for(g in 1:length(gene.lists)){
    file.name <- file.path(toa.cluster.dir, paste0("Cluster", g, "_", gsub(" ", "_", names(gene.lists)[g]), ".csv"))
    gene.names <- gene.info[match(gene.lists[[g]], gene.info[,"ensembl_gene_id"]),"external_gene_name"]
    gene.vals <- val.mat$value.mat[gene.lists[[g]],]
    cluster.table <- cbind(gene.lists[[g]], gene.names, gene.vals)
    colnames(cluster.table) <- c("gene.ID", "gene.name", tissue.names)
    write.table(cluster.table, file.name,
        sep = ",", quote = FALSE, row.names = FALSE, col.names = FALSE)
    full.list[[g]] <- cluster.table
}

#write them all out in one file
cluster.labels <- lapply(1:length(full.list), function(x) rep(alt.labels[x,1], nrow(full.list[[x]])))
full.table <- cbind("Cluster" = unlist(cluster.labels), Reduce("rbind", full.list))
#some gene names are missing. use gconvert to get all of them.
gene.names <- gconvert(full.table[,"gene.ID"], organism = "mmusculus")
#setdiff(full.table[,"gene.ID"], gene.names[,"target"])
common.genes <- intersect(full.table[,"gene.ID"], gene.names[,"target"])
table.idx <- match(common.genes, full.table[,"gene.ID"])
name.idx <- match(common.genes, gene.names[,"target"])
full.table[table.idx,"gene.name"] <- gene.names[name.idx,"name"]

write.table(full.table, file.path(toa.cluster.dir, "Tissues_of_Action_Genes.txt"), sep = "\t",
    quote = FALSE, row.names = FALSE)
```

```{r Fig6A}

#png(here("Results", "Paper", "Fig6A.png"), width = 6, height = 7, units = "in", res = 300)
pdf(here("Results", "Paper", "Fig6A.pdf"), width = 6.5, height = 7)
val.mat <- plot_clustered_values(cluster_results = toa.results, 
    enrichment.tables = FALSE, loading.viols = FALSE,
    alt.labels = alt.labels[,1], cluster.borders = "boxes")
dev.off()

fig.table <- cbind(rownames(val.mat[[1]]), val.mat$labels, val.mat[[1]])
colnames(fig.table)[1:2] <- c("geneID", "enrichment cluster")
write_table_with_header(fig.table, file.path(figure.data.dir, "Fig6A.txt"),
    header = "Data associated with Figure 6A. Table contains genes with the
    loadings greater than 2.5 standard deviations from the mean in any tissue.
    The columns indicate the gene ID, the enrichment cluster to which it 
    belonged, and the loading in each of the four tissues.")
```


```{r test_kegg, eval = FALSE}
all.kegg <- readRDS(here("Data", "general", "KEGG.Mouse.RDS"))
u_path <- gsub(" - Mus musculus (house mouse)", "", all.kegg[[2]][,"to"], fixed = TRUE)
path.id <- all.kegg[[2]][,1]
path.idx <- lapply(path.id, function(x) which(all.kegg[[1]][,1] == unlist(x)[1]))
path.gene.id <- lapply(path.idx, function(x) all.kegg[[1]][x,2])
names(path.gene.id) <- u_path

path.ensembl <- lapply(path.gene.id, function(x) gene.info[match(x, gene.info[,"entrezgene_id"]),"ensembl_gene_id"])

path.kegg.overlap <- matrix(NA, nrow = length(path.ensembl), ncol = length(gene.lists))
rownames(path.kegg.overlap) <- names(path.ensembl)
colnames(path.kegg.overlap) <- names(gene.lists)
for(i in 1:length(path.ensembl)){
    for(j in 1:length(gene.lists)){
        path.kegg.overlap[i,j] <- length(intersect(path.ensembl[[i]], gene.lists[[j]]))
    }
}
pdf("~/Desktop/test_kegg.pdf", width = 12, height = 40)
pheatmap(path.kegg.overlap, display_numbers = TRUE)
dev.off()

path.name <- "Apelin signaling pathway"
path.name <- "Adipocytokine signaling pathway"
path.name <- "Insulin resistance"
path.name <- "Oxidative phosphorylation"
path.name <- "Fatty liver disease"
pathway.idx <- which(u_path == path.name)
path.genes <- path.ensembl[[pathway.idx]]
path.in.top <- intersect(path.genes, rownames(val.mat))
path.gene.names <- gene.info[match(path.in.top, gene.info[,"ensembl_gene_id"]),"external_gene_name"]
path.mat <- val.mat[path.in.top,]
rownames(path.mat) <- path.gene.names
pheatmap(path.mat)
boxplot(path.mat);abline(h = 0)
```

```{r check_loading, eval = FALSE}
top.genes <- gene.info[match(rownames(val.mat), gene.info[,"ensembl_gene_id"]),"external_gene_name"]
sort(top.genes)

gene.name <- "Lepr"
gene.id <- gene.info[which(gene.info[,"external_gene_name"] == gene.name),"ensembl_gene_id"]
gene.loadings <- unlist(sapply(transcript_loadings, function(x) x[which(rownames(x) == gene.id)]))
barplot(gene.loadings)
```

```{r fig6B}

gene.name <- "Pparg"
gene.id <- gene.info[which(gene.info[,"external_gene_name"] == gene.name),"ensembl_gene_id"]
gene.loadings <- unlist(sapply(transcript_loadings, function(x) x[which(rownames(x) == gene.id)]))

pdf(here("Results", "Paper", "Fig6B.pdf"), width = 6, height = 5.5)
barplot_with_num(signif(gene.loadings, 2), col = tissue.cols, ylab = "Loading",
    main = gene.name)
dev.off()

fig.table <- matrix(signif(gene.loadings, 2), nrow = 1)
colnames(fig.table) <- names(gene.loadings)
write_table_with_header(fig.table, file.path(figure.data.dir, "Fig6B.txt"),
    header = "Data associated with Figure 6B. Loadings of the gene Pparg
    in each of the four tissues. Numbers rounded to two significant digits.")
```


# Extra

## eQTL predictions in DO

Can we predict phenotype in the DO using loadings and locally
imputed gene expression? 

```{r pred}
local.impute <- list.files(here("Results", "DO", "Imputed_Transcriptomes"), 
    pattern = "Local_Imputation")
```

## Loadings vs. LOD {.tabset .tabset-fade .tabset-pills}

I don't think this is relevant anymore, but I don't want to
take the code out completely.

Another way to look at the relationship between transcript
loading and local variance explained is to use LOD score
as a measure of the variance explained. Geneticists are 
also used to looking at eQTL LOD scores.

What is the maximum eQTL LOD score for transcripts compared with 
their loading? The following plots show the loading of each 
transript vs. its maximum LOD score. We see a negative
relationship between loading and LOD score.

```{r to_label}
num.label = 30 #the number of top genes to label in the plots below
```

The `r num.label` genes along the edge are labeled.

The plots are saved as pdfs called LOD_vs_Loading_tissue_name
so that the gene names are searchable and highlightable.

```{r lod_v_loading, fig.width = 8, fig.height = 6, results = "asis"}
convex.layers <- 3

if(mediation.text == "complete_mediation"){
    for(tx in 1:length(tissue.names)){
        cat("####", tissue.names[tx], "\n")

        eqtl.table <- tissue.lod[[tx]]

        gene.id <- rownames(transcript_loadings[[tx]])
        gene.name <- gene.tables[[tx]]$symbol[match(gene.id, gene.tables[[tx]]$gene.id)]
        gene.lod <- lapply(gene.id, function(x) eqtl.table$lod[which(eqtl.table$gene.id == x)])
        max.lod <- sapply(gene.lod, function(x) if(length(x) > 0){max(x)}else{3})
        xmax <- max(max.lod)*1.1
                
        #pdf(file.path(results.dir, paste0("LOD_vs_Loading_", tissue.names[tx], "_", adjust_numeral, ".pdf")), width = 10, height = 8)
        layout(matrix(c(1,1,2,3), ncol = 2, byrow = FALSE), widths = c(1, 0.5))
        par(mar = c(4,4,4,0))

        convex.idx <- vector(mode = "list", length = convex.layers)
        lod.idx <- loading.idx <- 1:length(max.lod)        
        x <- max.lod[lod.idx]
        y <- transcript_loadings[[tx]][lod.idx,]
        for(cv in 1:convex.layers){
            hull <- convex_hull(cbind(x,y))
            convex.idx[[cv]] <- names(y)[hull[[1]]]
            x <- x[-hull[[1]]]
            y <- y[-hull[[1]]]
        }
        hull.idx <- match(unlist(convex.idx), rownames(transcript_loadings[[tx]]))
        #don't label anything with a local LOD score < 25
        high.lod <- which(max.lod > 25)
        hull.idx <- intersect(hull.idx, high.lod)
        plot(max.lod, transcript_loadings[[tx]], xlab = "Maximum LOD", ylab = "Loading", 
            main = tissue.names[tx], pch = 16, cex = 0.5, xlim = c(0, xmax))
        points(max.lod[hull.idx], transcript_loadings[[tx]][hull.idx,1], col = "red")
        text(max.lod[hull.idx], transcript_loadings[[tx]][hull.idx,1], 
            labels = gene.name[hull.idx], pos = 4, cex = 0.7)
        abline(h = 0)  
        
        pos.hull.idx <- intersect(hull.idx, which(transcript_loadings[[tx]][,1] > 0))
        top.genes <- gene.name[pos.hull.idx]
        ordered.top <- top.genes[order(max.lod[pos.hull.idx], decreasing = TRUE)]
        pos.par <- words_to_paragraph(ordered.top, line.len = 5)
        par(mar = c(0,0,0,0))
        plot.text(paste("Positive Loading Genes", pos.par, sep = "\n\n"))
        
        neg.hull.idx <- intersect(hull.idx, which(transcript_loadings[[tx]][,1] < 0))
        bottom.genes <- gene.name[neg.hull.idx]
        ordered.bottom <- bottom.genes[order(max.lod[neg.hull.idx], decreasing = TRUE)]
        neg.par <- words_to_paragraph(ordered.bottom, line.len = 5)
        par(mar = c(0,0,0,0))
        plot.text(paste("Negative Loading Genes", neg.par, sep = "\n\n"))
        #dev.off()
        cat("\n\n")
    }
}

```

### Variance vs. LOD

Local eQTLs tend to be concordant across tissues, whereas
distal eQTLs do not tend to match. Is there a relationship
between loading variance across tissues and LOD scores?
We would expect that genes that need to be differentially
regulated across tissues would have high variance in loading
across tissues, and low local heritability. 

```{r var_v_lod, fig.width = 8, fig.height = 8}
common.tx <- Reduce("intersect", lapply(transcript_loadings, rownames))
common.tx.mat <- sapply(transcript_loadings, function(x) x[common.tx,1])
cross.tissue.var <- apply(common.tx.mat, 1, var)


#hist(cross.tissue.var)

par(mfrow = c(2,2))
for(tx in 1:length(tissue.names)){
    eqtl.table <- tissue.lod[[tx]]
    gene.lod <- lapply(common.tx, function(x) eqtl.table$lod[which(eqtl.table$gene.id == x)])
    max.lod <- sapply(gene.lod, function(x) if(length(x) > 0){max(x)}else{3})
    plot.new()
    plot.window(xlim = c(0, max(cross.tissue.var)), ylim = c(0, 250))
    points(cross.tissue.var, max.lod, col = "gray10")
    plot.with.model(cross.tissue.var, max.lod, report = "cor.test", add = TRUE,
        col = tissue.cols[tx], line.col = "gray50")
    axis(1);axis(2)
    mtext(tissue.names[tx], side = 3)
    mtext("Cross-Tissue Variance", side = 1, line = 2.5)
    mtext("Maximum LOD score", side = 2, line = 2.5)
}
```

Are transcripts important in individual tissues? Two tissues? More
Rank Z transform the transcript loadings in each tissue, then count
the number of tissues that each transcript has a large loading (>= 2.5) 
in.

The barplot below shows that for most transcripts that have a large
loading, they have that large loading in only a single tissue.

```{r importance}
z.thresh <- 2.5
norm.loadings <- apply(common.tx.mat, 2, rankZ)
num.tx <- apply(norm.loadings, 1, function(x) length(which(abs(x) >= z.thresh)))
num.count <- sapply(1:4, function(x) length(which(num.tx == x)))

barplot_with_num(num.count, ylab = "transcript count", xlab = "Number of Tissues")
```


```{r tx_pairs}
tx.pairs <- which(num.tx > 0)
pheatmap(common.tx.mat[names(tx.pairs),], show_rownames = FALSE, cluster_cols = FALSE)
tx.pair.genes <- gene.info[match(names(tx.pairs), gene.info[,1]),]
pair.enrich <- gost(names(tx.pairs), organism = "mmusculus")
plot.enrichment(pair.enrich)
```

Taken together, these results suggest that transcripts that are important
for traits are differentially regulated across tissues, and thus cannot
have strong local eQTL. 


## Compare Paths

One of the reviewers of the manuscript has asked about the
strength of the composite score correlations given that there
are eQTLs in the transcriptome. They thought the permutations
did not address whether the transcriptome-phenome correlations
were particularly strong given that there are eQTLs. 

To better address this question, we have performed HDMA again 
with two more transcriptome kernels: one derived from the 
residual expression after local haplotype effects have been 
removed (the distal-only kernen), and the other derived from
the locally imputed transcripts (local-only kernel).

The local-only kernel resulted in an almost perfect G-T
correlation (0.92) because the transcriptome is based
purely on the genome. It resulted in a very low T-P
correlation (0.14) because the transcripts with large 
local eQTLs are heavily weighted in this transcriptome,
and they are not well correlated with traits. It says, 
though, that if you share a lot of local eQTL with another 
mouse, you will have similar traits to some extent.

The distal-only kernel gave results that were extremely
similar to the full transcriptome. The G-T correlation
was slightly lower, which makes sense, because we removed
some of the direct genetic effect from the transcriptome.
The T-P correlation was still very high, suggesting that
the T-P correlation is the result of distally inherited
variation, not locally inherited variation. 

The local-only kernel also does not give us a significant
path coefficient.

Here we plot the results of these kernels on the permutation
plot in Figure 3C. 

```{r compare_paths}
#All use.local/use.distal combinations need to be run
#before running this code. Make sure that the FALSE/FALSE
#situation was the last one run so that all the figures
#use the correct data.

box_text <- function(center.point.x, center.point.y, width, height, 
    fill = "white", border = "black", lwd = 1){
    draw.rectangle(center.point.x-width, center.point.x+width, 
        center.point.y-height, center.point.y+height, fill = fill, 
        border.col = border, lwd = lwd)
}

distal.dir <- gsub("High_Dim_Med", "High_Dim_Med_distal", results.dir)
local.dir <- gsub("High_Dim_Med", "High_Dim_Med_local", results.dir)

if(file.exists(distal.dir) && file.exists(local.dir)){
    
    tx = 1

    local.path.stats <- readRDS(file.path(local.dir, "Path_Stats.RDS"))
    distal.path.stats <- readRDS(file.path(distal.dir, "Path_Stats.RDS"))

    all.stats <- list("all" = t(as.matrix(tissue.path.stats)), 
        "local" = t(as.matrix(local.path.stats)), 
        "distal" = t(as.matrix(distal.path.stats)),
        "perm" = all.perms[[tx]])

    pdf(here("Results", "Paper", "Fig3B_C.pdf"), width = 5, height = 5)

    tissue.path.coef <- tissue.path.stats[1]

    axis.max <- max(unlist(lapply(all.stats, function(x) x[,c("X-M", "M-Y")])))
    axis.min <- min(unlist(lapply(all.stats, function(x) x[,c("X-M", "M-Y")])))
    #y.axis.max <- max(unlist(lapply(all.stats, function(x) x[,"M-Y"])))
    #y.axis.min <- min(unlist(lapply(all.stats, function(x) x[,"M-Y"])))

    text.size = 1.5
    arrow.len = 0.15

    plot(all.perms[[tx]][,"X-M"], all.perms[[tx]][,"M-Y"], 
        xlim = c(axis.min, axis.max), 
        ylim = c(axis.min, axis.max), 
        xlab = "", ylab = "")
    mtext(bquote(italic(G[C])~"-"~italic(T[C])~" correlation ("*italic(r)~")"), 
        side = 1, line = 2.8, cex = text.size)
    mtext(bquote(italic(T[C])~"-"~italic(P[C])~" correlation ("*italic(r)~")"), 
        side = 2, line = 2.5, cex = text.size)
    abline(0,1)

    text(x = 0.35, y = 0.2, labels = "Null", cex = text.size)
    arrows(x0 = 0.42, x1 = 0.52, y0 = 0.2, length = arrow.len, lwd = 3)
    
    #local model
    points(local.path.stats["X-M"], local.path.stats["M-Y"], 
        col = local.col, pch = 16, cex = 2)
    text(x = 0.78, y = 0.35, labels = "Local\nModel", cex = text.size)
    arrows(x0 = 0.78, x1 = local.path.stats["X-M"]-0.025, 
        y0 = 0.27, y1 = local.path.stats["M-Y"]+0.01, lwd = 3, length = 0.1)

    #distal model
    points(distal.path.stats["X-M"], distal.path.stats["M-Y"], 
        col = distal.col, pch = 16, cex = 2)    
    text(x = 0.18, y = 0.6, labels = "Distal\nModel", cex = text.size)
    arrows(x0 = 0.27, x1 = distal.path.stats["X-M"]-0.02, 
        y0 = 0.68, y1 = distal.path.stats["M-Y"]-0.007, lwd = 3, length = 0.1)
    
    #full model
    points(tissue.path.stats["X-M"], tissue.path.stats["M-Y"], 
        col = "red", pch = 16, cex = 2)
    box_text(0.7, 0.8, 0.15, 0.05, border = NA)
    text(x = 0.7, 0.8, "Observed", cex = text.size)
    arrows(x0 = 0.55, x1 = tissue.path.stats["X-M"]+0.025, y0 = 0.8, 
        y1 = tissue.path.stats["M-Y"]+0.005, lwd = 3, length = 0.1)

    #=============================================
    #density plot of null path coefficient
    #=============================================

    local.path <- local.path.stats["path.coef"]
    distal.path <- distal.path.stats["path.coef"]
    
    path.min <- min(c(all.perms[[tx]][,"path_coef"], tissue.path.coef), na.rm = TRUE)
    #path.max <- max(c(all.perms[[tx]][,"path_coef"], tissue.path.coef), na.rm = TRUE)
    path.max = 0.40

    text.size = 1.5

    #hist(all.perms[[tx]][,"path_coef"], xlim = c(path.min, path.max), main = "", breaks = 100, xlab = "", ylab = "")
    plot(density(all.perms[[tx]][,"path_coef"]), lwd = 3,
        xlim = c(path.min, path.max), main = "", xlab = "", ylab = "")
    plot.dim <- par("usr")
    mtext("Path Coefficient", side = 1, line = 2.5, cex = text.size)
    mtext("Frequency", side = 2, line = 2.5, cex = text.size)

    plot.height = plot.dim[4] - plot.dim[3]
    observed.height = plot.height*0.5
    other.height = plot.height**0.3
    box.height = plot.height*0.05
    box.lwd = 2

    #box_text(0.2, plot.height*0.85, 0.04, box.height, lwd = box.lwd)
    text(0.2, plot.height*0.85, labels = "Null", cex = text.size)
    arrows(x0 = 0.17, x1 = 0.13, y0 = plot.height*0.85, 
        y1 = (plot.height*0.85)-1, length = arrow.len, lwd = 3)

    arrows(x0 = tissue.path.coef, y0 = observed.height, y1 = 0, 
        col = "red", lwd = 3, length = arrow.len)
    box_text(tissue.path.coef, observed.height+1, 0.08, box.height, 
        border = "red", lwd = box.lwd)
    text(x = tissue.path.coef, y = observed.height+1, labels = "Observed",
        cex = text.size)
    
    #arrows(x0 = local.path, y0 = other.height, y1 = 0, col = "black", lwd = 5, length = arrow.len) 
    arrows(x0 = local.path, y0 = other.height, y1 = 0, col = local.col, lwd = 3,
        length = arrow.len)
    box_text(local.path, other.height+1, 0.09, box.height, 
        border = local.col, lwd = box.lwd) 
    text(x = local.path, y = other.height+1, labels = "Local Model",
        cex = text.size)

    arrows(x0 = distal.path, y0 = other.height, y1 = 0, col = distal.col, 
        lwd = 3, length = arrow.len)
    box_text(distal.path, other.height+1, 0.09, box.height, 
        border = distal.col, lwd = box.lwd) 
    text(x = distal.path, y = other.height+1, labels = "Distal Model",
        cex = text.size)

    dev.off()
}


write_table_with_header(all.perms[[1]][,1,drop=FALSE], file.path(figure.data.dir,
    "Fig3B_null.txt"), header = "Data corresponding to Figure 3B 
    null distribution. Null path coefficients for 1000 permutations.")

fig.table <- matrix(c(tissue.path.coef, local.path, distal.path),  nrow = 1)
colnames(fig.table) <- c("Observed", "Local", "Distal")
write_table_with_header(fig.table, file.path(figure.data.dir, "Fig3B_stats.txt"),
    header = "Data corresponding to arrows in Figure 3B. Observed 
    path coefficients for full model, model with local genetic 
    effects only, and model with distal genetic effects only.")

fig.table <- all.stats$perm[,c("X-M", "M-Y")]
colnames(fig.table) <- c("G-T correlation", "T-P correlation")
write_table_with_header(fig.table, file.path(figure.data.dir, "Fig3C_null.txt"),
    header = "Data corresponding to null distribution in Figure 3C. 
    Null distribution of G-T correlation and T-P correlation from 1000 
    permutations.")


fig.table <- t(sapply(all.stats[1:3], function(x) x[,c("X-M", "M-Y")]))
fig.table <- cbind(rownames(fig.table), fig.table)
colnames(fig.table) <- c("Genetic Contributions", "G-T", "T-P")
write_table_with_header(fig.table, file.path(figure.data.dir, "Fig3C_stats.txt"),
    header = "Data corresponding to observed 
    statistics in Figure 3C. G-T and T-P correlations for models 
    including all genetic effects, local genetic effects only, and 
    distal genetic effects only.")
```


## Nucb2

Nucb2 has high distal heritability in the islet.
Perform a qtl scan to see if it has an significant
distal eQTL.

```{r nucb2, fig.width = 10, fig.height = 3.5}
gene.name = "Nucb2"; tissue.name = "Islet"

gene.idx <- which(gene.info$external_gene_name == gene.name)
gene.id <- gene.info$ensembl_gene_id[gene.idx]
gene.chr <- gene.info$chromosome_name[gene.idx]
gene.pos <- gene.info$start_position[gene.idx]/1e6
tissue.idx <- which(names(scaled.expr) == tissue.name)
common.ind <- Reduce("intersect", lapply(Kt, colnames))
gene.expr <- adj.expr[[tissue.idx]][common.ind,gene.id]
gene.scan <- scan1(genoprobs, gene.expr, addcovar = covar, kinship = Kg)
plot(gene.scan, map = map, main = paste(gene.name, "in", tissue.name))
```


```{r nucb2_fig}
pdf(here("Results", "Paper", "Supp_Fig_14_Nucb2.pdf"), width = 10, 
    height = 3.5)
par(mar = c(2,4,2,0))
plot_scan1(gene.scan, map, lodcol = 1, ylim = c(0, 9), 
    chr = c(1:19, "X"), main = substitute(paste(italic("Nucb2"), " in islet")), 
    type = "l", ylab = "LOD score",
    col = "gray20", mark.chr = gene.chr, mark.pos = gene.pos, mark.col = "red", 
    mark.lwd = 2, mark.type = "arrow", arrow.len = 0.1, arrow.max = 6,
    mark.label = "Nucb2", mark.font = 3)
plot.dim <- par("usr")
segments(x0 = plot.dim[1], x1 = plot.dim[2]*0.95, y0 = 8, lty = 2)
dev.off()

fig.table <- data.frame("marker_name" = rownames(gene.scan),
    "lod" = gene.scan[,1])
write_table_with_header(fig.table, file.path(figure.data.dir, "Supp_Fig14.txt"),
    header = "Data corresponding to Supplementary Figure 14. Table includes
    all marker names, which are formatted with the chromosome name and
    the marker position in base pairs separated by an underscore.
    It also includes the LOD score
    measured at each marker for the transcript level of Nucb2 in pancreatic
    islet in the DO. Nucb2 is encoded on chromosome 7 at 116.5Mb.")
```
